% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_ds_to_dv_repo}
\alias{add_ds_to_dv_repo}
\title{add dataset to dataverse repository}
\usage{
add_ds_to_dv_repo(
  dv_1L_chr,
  ds_meta_ls,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{dv_1L_chr}{Dataverse (a character vector of length one)}

\item{ds_meta_ls}{Dataset meta (a list)}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Dataset url (a character vector of length one)
}
\description{
add_ds_to_dv_repo() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dataset to dataverse repository. Function argument dv_1L_chr specifies the object to be updated. The function returns Dataset url (a character vector of length one).
}
