% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_dv_meta_to_imp_lup}
\alias{add_dv_meta_to_imp_lup}
\title{Add dataverse meta to import}
\usage{
add_dv_meta_to_imp_lup(
  imp_lup,
  ds_ui_1L_chr,
  file_type_1L_chr,
  save_type_1L_chr
)
}
\arguments{
\item{imp_lup}{Import (a lookup table)}

\item{ds_ui_1L_chr}{Dataset ui (a character vector of length one)}

\item{file_type_1L_chr}{File type (a character vector of length one)}

\item{save_type_1L_chr}{Save type (a character vector of length one)}
}
\value{
Import (a lookup table)
}
\description{
add_dv_meta_to_imp_lup() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add dataverse meta to import lookup table. Function argument imp_lup specifies the object to be updated. The function returns Import (a lookup table).
}
\keyword{internal}
