% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_fields_from_lup}
\alias{add_fields_from_lup}
\title{Add fields from lookup table}
\usage{
add_fields_from_lup(ds_tb, lup_tb, match_chr, target_1L_chr, vars_chr)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{lup_tb}{Lookup table (a tibble)}

\item{match_chr}{Match (a character vector)}

\item{target_1L_chr}{Target (a character vector of length one)}

\item{vars_chr}{Variables (a character vector)}
}
\value{
Dataset (a tibble)
}
\description{
add_fields_from_lup() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add fields from lookup table. The function returns Dataset (a tibble).
}
\keyword{internal}
