% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_files_to_dv}
\alias{add_files_to_dv}
\title{add files to dataverse}
\usage{
add_files_to_dv(
  files_tb,
  data_dir_rt_1L_chr = ".",
  ds_url_1L_chr,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{files_tb}{Files (a tibble)}

\item{data_dir_rt_1L_chr}{Data directory root (a character vector of length one), Default: '.'}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
File identities (an integer vector)
}
\description{
add_files_to_dv() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add files to dataverse. Function argument files_tb specifies the object to be updated. The function returns File identities (an integer vector).
}
\keyword{internal}
