% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_from_lup_prototype}
\alias{add_from_lup_prototype}
\title{Add from lookup table prototype}
\usage{
add_from_lup_prototype(
  data_tb,
  arrange_1L_chr = character(0),
  exclude_chr = character(0),
  lup_prototype_tb = NULL,
  match_var_nm_1L_chr = "UID_chr",
  method_1L_chr = c("first", "sample"),
  type_1L_chr = c("sequential", "batch", "self"),
  vars_chr = character(0)
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{arrange_1L_chr}{Arrange (a character vector of length one), Default: character(0)}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{lup_prototype_tb}{Lookup table prototype (a tibble), Default: NULL}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: 'UID_chr'}

\item{method_1L_chr}{Method (a character vector of length one), Default: c("first", "sample")}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("sequential", "batch", "self")}

\item{vars_chr}{Variables (a character vector), Default: character(0)}
}
\value{
Data (a tibble)
}
\description{
add_from_lup_prototype() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add from lookup table prototype. The function returns Data (a tibble).
}
