% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_latest_match}
\alias{add_latest_match}
\title{Add latest match}
\usage{
add_latest_match(
  data_tb,
  dynamic_lup,
  target_var_nm_1L_chr,
  date_var_1L_chr = "Date",
  invert_1L_lgl = FALSE,
  match_var_nm_1L_chr = "UID_chr",
  type_1L_chr = c("chr", "dbl", "int", "lgl")
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{dynamic_lup}{Dynamic (a lookup table)}

\item{target_var_nm_1L_chr}{Target variable name (a character vector of length one)}

\item{date_var_1L_chr}{Date variable (a character vector of length one), Default: 'Date'}

\item{invert_1L_lgl}{Invert (a logical vector of length one), Default: FALSE}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: 'UID_chr'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("chr", "dbl", "int", "lgl")}
}
\value{
Data (a tibble)
}
\description{
add_latest_match() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add latest match. The function returns Data (a tibble).
}
