% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_significance}
\alias{add_significance}
\title{Add significance}
\usage{
add_significance(
  plot_plt,
  by_1L_chr,
  data_tb,
  var_1L_chr,
  add_1L_dbl = numeric(0),
  adjust_1L_dbl = 0.4,
  as_percent_1L_lgl = F,
  digits_1L_int = 4,
  flip_1L_lgl = F,
  scientific_1L_lgl = F,
  show_p_1L_lgl = T,
  show_test_1L_lgl = F,
  tip_1L_dbl = 0,
  ...
)
}
\arguments{
\item{plot_plt}{Plot (a plot)}

\item{by_1L_chr}{By (a character vector of length one)}

\item{data_tb}{Data (a tibble)}

\item{var_1L_chr}{Variable (a character vector of length one)}

\item{add_1L_dbl}{Add (a double vector of length one), Default: numeric(0)}

\item{adjust_1L_dbl}{Adjust (a double vector of length one), Default: 0.4}

\item{as_percent_1L_lgl}{As percent (a logical vector of length one), Default: F}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 4}

\item{flip_1L_lgl}{Flip (a logical vector of length one), Default: F}

\item{scientific_1L_lgl}{Scientific (a logical vector of length one), Default: F}

\item{show_p_1L_lgl}{Show p (a logical vector of length one), Default: T}

\item{show_test_1L_lgl}{Show test (a logical vector of length one), Default: F}

\item{tip_1L_dbl}{Tip (a double vector of length one), Default: 0}

\item{...}{Additional arguments}
}
\value{
Plot (a plot)
}
\description{
add_significance() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add significance. The function returns Plot (a plot).
}
\keyword{internal}
