% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_with_join}
\alias{add_with_join}
\title{Add with join}
\usage{
add_with_join(X_Ready4useDyad, Y_Ready4useDyad)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION}

\item{Y_Ready4useDyad}{PARAM_DESCRIPTION}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
add_with_join() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add with join. The function is called for its side effects and does not return a value.
}
