% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_assert.R
\name{assert_file_exists}
\alias{assert_file_exists}
\title{Assert file exists}
\usage{
assert_file_exists(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\description{
assert_file_exists() is an Assert function that validates that an object conforms to required condition(s). If the object does not meet all required conditions, program execution will be stopped and an error message provided. Specifically, this function implements an algorithm to assert a file exists. Function argument x specifies the object on which assert validation checks are to be performed. Argument NA provides the object containing values used for validation tests.NA
}
\keyword{internal}
