% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_author.R
\name{author.ready4use_manifest}
\alias{author.ready4use_manifest}
\alias{author,ready4use_manifest-method}
\title{Author method applied to ready4 S3 class defining a manifest of data required to create an package..}
\usage{
author.ready4use_manifest(x)

\S4method{author}{ready4use_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 s3 class defining a manifest of data required to create an R package.}
}
\value{
x (An object)
}
\description{
author.ready4use_manifest() is an Author method that writes files to local or remote locations. This method is implemented for the ready4 s3 class defining a manifest of data required to create an R package. The function is called for its side effects and does not return a value.
}
