% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_authorData.R
\name{authorData.ready4use_manifest}
\alias{authorData.ready4use_manifest}
\alias{authorData,ready4use_manifest-method}
\title{AuthorData - a method that authors and saves files necessary for creating and documenting datasets.}
\usage{
authorData.ready4use_manifest(x)

\S4method{authorData}{ready4use_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 s3 class defining a manifest of data required to create an R package.}
}
\description{
authorData.ready4use_manifest() is an AuthorData method that authors and saves files necessary for creating and documenting datasets. This method is implemented for the ready4 s3 class defining a manifest of data required to create an R package. The function is called for its side effects and does not return a value.
}
