% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_bind.R
\name{bind_dyads}
\alias{bind_dyads}
\title{Bind dyads}
\usage{
bind_dyads(
  dyad_ls,
  drop_chr = character(0),
  factors_chr = character(0),
  tfmn_fn = identity,
  uid_var_nm_1L_chr = character(0)
)
}
\arguments{
\item{dyad_ls}{Dyad (a list)}

\item{drop_chr}{Drop (a character vector), Default: character(0)}

\item{factors_chr}{Factors (a character vector), Default: character(0)}

\item{tfmn_fn}{Transformation (a function), Default: identity}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: character(0)}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
bind_dyads() is a Bind function that binds two objects together to create a composite object. Specifically, this function implements an algorithm to bind dyads. The function is called for its side effects and does not return a value.
}
\keyword{internal}
