% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{bind_lups}
\alias{bind_lups}
\title{Bind lookup tables}
\usage{
bind_lups(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
bind_lups() is a Bind Lookup Tables generic that rowbinds lookup tables of the same class, removing duplicates based on priority.
}
