% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_crs_nbr_vec.R
\name{crs_nbr_vec}
\alias{crs_nbr_vec}
\alias{crs_nbr_vec-ready4_script_data}
\title{crs_nbr_vec}
\usage{
crs_nbr_vec(x)

\S4method{crs_nbr_vec}{ready4_script_data}(x)
}
\arguments{
\item{x}{An object of class ready4_script_data}
}
\description{
S4 Generic function to get the value of the slot crs_nbr_vec

Get the value of the slot crs_nbr_vec for S4 objects of class ready4_script_data
}
