% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_depict.R
\name{depict-Ready4useDyad}
\alias{depict-Ready4useDyad}
\alias{depict,Ready4useDyad-method}
\title{Depict (plot) features of model module data}
\usage{
\S4method{depict}{Ready4useDyad}(
  x,
  x_vars_chr = character(0),
  y_vars_chr = character(0),
  z_vars_chr = character(0),
  arrange_1L_lgl = FALSE,
  arrange_args_ls = list(),
  as_percent_1L_lgl = FALSE,
  colours_chr = c("#de2d26", "#fc9272"),
  drop_legend_1L_lgl = FALSE,
  drop_missing_1L_lgl = FALSE,
  drop_ticks_1L_lgl = FALSE,
  fill_single_1L_lgl = FALSE,
  line_1L_chr = "black",
  position_xx = NULL,
  recode_lup_r3 = ready4show::ready4show_correspondences(),
  significance_1L_lgl = F,
  significance_args_ls = list(),
  style_1L_chr = get_styles(),
  titles_chr = character(0),
  type_1L_chr = c("ggsci", "manual", "viridis"),
  x_labels_chr = character(0),
  y_labels_chr = character(0),
  z_labels_chr = character(0),
  what_1L_chr = get_journal_plot_fn("names"),
  ...
)
}
\arguments{
\item{x}{An object of class Ready4useDyad}

\item{x_vars_chr}{X variables (a character vector), Default: character(0)}

\item{y_vars_chr}{Y variables (a character vector), Default: character(0)}

\item{z_vars_chr}{Z variables (a character vector), Default: character(0)}

\item{arrange_1L_lgl}{Arrange (a logical vector of length one), Default: FALSE}

\item{arrange_args_ls}{Arrange arguments (a list), Default: list()}

\item{as_percent_1L_lgl}{As percent (a logical vector of length one), Default: FALSE}

\item{colours_chr}{Colours (a character vector), Default: c("#de2d26", "#fc9272")}

\item{drop_legend_1L_lgl}{Drop legend (a logical vector of length one), Default: FALSE}

\item{drop_missing_1L_lgl}{Drop missing (a logical vector of length one), Default: FALSE}

\item{drop_ticks_1L_lgl}{Drop ticks (a logical vector of length one), Default: FALSE}

\item{fill_single_1L_lgl}{Fill single (a logical vector of length one), Default: FALSE}

\item{line_1L_chr}{Line (a character vector of length one), Default: 'black'}

\item{position_xx}{Position (an output object of multiple potential types), Default: NULL}

\item{recode_lup_r3}{Recode (a ready4 submodule extension of lookup table), Default: ready4show::ready4show_correspondences()}

\item{significance_1L_lgl}{Significance (a logical vector of length one), Default: F}

\item{significance_args_ls}{Significance arguments (a list), Default: list()}

\item{style_1L_chr}{Style (a character vector of length one), Default: get_styles()}

\item{titles_chr}{Titles (a character vector), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("ggsci", "manual", "viridis")}

\item{x_labels_chr}{X labels (a character vector), Default: character(0)}

\item{y_labels_chr}{Y labels (a character vector), Default: character(0)}

\item{z_labels_chr}{Z labels (a character vector), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: get_journal_plot_fn("names")}

\item{...}{Additional arguments}
}
\value{
Plot (an output object of multiple potential types)
}
\description{
depict method applied to Ready4useDyad
}
