% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{download_data}
\alias{download_data}
\title{Download data}
\usage{
download_data(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
download_data() is a Download Data generic that downloads data files.NA
}
\keyword{internal}
