% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit-Ready4useDyad}
\alias{exhibit-Ready4useDyad}
\alias{exhibit,Ready4useDyad-method}
\title{exhibit}
\usage{
\S4method{exhibit}{Ready4useDyad}(
  x,
  caption_1L_chr = NA_character_,
  display_1L_chr = "all",
  mkdn_tbl_ref_1L_chr = "",
  output_type_1L_chr = "HTML",
  type_1L_chr = "ds",
  use_lbls_as_col_nms_1L_lgl = T,
  use_rdocx_1L_lgl = F,
  ...
)
}
\description{
exhibit method applied to Ready4useDyad
}
