% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_fns_dmt_tb.R
\docType{data}
\name{fns_dmt_tb}
\alias{fns_dmt_tb}
\title{ready4use function documentation table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 10 columns.
}
\source{
\url{https://ready4-dev.github.io/ready4/}
}
\usage{
fns_dmt_tb
}
\description{
A table with the summary information on functions included in the ready4use package.
}
\details{
A tibble

\describe{
\item{fns_chr}{Functions (a character vector)}
\item{title_chr}{Title (a character vector)}
\item{desc_chr}{Description (a character vector)}
\item{details_chr}{Details (a character vector)}
\item{inc_for_main_user_lgl}{Include for main user (a logical vector)}
\item{output_chr}{Output (a character vector)}
\item{example_lgl}{Example (a logical vector)}
\item{args_ls}{Arguments (a list)}
\item{file_nm_chr}{File name (a character vector)}
\item{file_pfx_chr}{File prefix (a character vector)}
}
}
\keyword{datasets}
