% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_colour_codes}
\alias{get_colour_codes}
\title{Get colour codes}
\usage{
get_colour_codes(
  colour_1L_int = 1,
  manual_chr = c("#de2d26", "#fc9272"),
  pick_1L_int = integer(0),
  single_1L_lgl = FALSE,
  strict_1L_lgl = FALSE,
  style_1L_chr = get_styles(),
  type_1L_chr = c("ggsci", "manual", "unicol", "viridis"),
  validate_1L_lgl = TRUE
)
}
\arguments{
\item{colour_1L_int}{Colour (an integer vector of length one), Default: 1}

\item{manual_chr}{Manual (a character vector), Default: c("#de2d26", "#fc9272")}

\item{pick_1L_int}{Pick (an integer vector of length one), Default: integer(0)}

\item{single_1L_lgl}{Single (a logical vector of length one), Default: FALSE}

\item{strict_1L_lgl}{Strict (a logical vector of length one), Default: FALSE}

\item{style_1L_chr}{Style (a character vector of length one), Default: get_styles()}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("ggsci", "manual", "unicol", "viridis")}

\item{validate_1L_lgl}{Validate (a logical vector of length one), Default: TRUE}
}
\value{
Colour codes (a character vector)
}
\description{
get_colour_codes() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get colour codes. The function returns Colour codes (a character vector).
}
\keyword{internal}
