% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_get_data.R
\name{get_data}
\alias{get_data}
\alias{get_data.ready4_dv_import_lup}
\alias{get_data,ready4_dv_import_lup-method}
\title{Get data}
\usage{
get_data(x, ...)

get_data.ready4_dv_import_lup(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

\S4method{get_data}{ready4_dv_import_lup}(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.}

\item{...}{Additional arguments}

\item{save_dir_path_1L_chr}{Save directory path (a character vector of length one), Default: ''}

\item{unlink_1L_lgl}{Unlink (a logical vector of length one), Default: T}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}
}
\description{
get_data() is a Get Data generic that retrieves data from R objects loaded in memory.

get_data.ready4_dv_import_lup() is a Get Data method that retrieves data from R objects loaded in memory. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.. The function is called for its side effects and does not return a value.
}
