% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
get_data() is a Get Data generic that retrieves data from R objects loaded in memory.
}
