% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_data.R
\name{get_data.ready4_dv_import_lup}
\alias{get_data.ready4_dv_import_lup}
\title{Get data.ready4 dataverse import}
\usage{
\method{get_data}{ready4_dv_import_lup}(x, save_dir_path_chr = "", unlink_lgl = T)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{save_dir_path_chr}{Save directory path (a character vector of length 1), Default: ''}

\item{unlink_lgl}{Unlink (a logical vector of length 1), Default: T}
}
\description{
get_data.ready4_dv_import_lup() is a Get Data generic that retrieves data from R objects loaded in memory.NA
}
\keyword{internal}
