% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_drop_offs}
\alias{get_drop_offs}
\title{Get drop offs}
\usage{
get_drop_offs(
  X_Ready4useDyad = Ready4useDyad(),
  condition_1L_chr = ">1",
  uid_var_nm_1L_chr = "uid_chr"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: Ready4useDyad()}

\item{condition_1L_chr}{Condition (a character vector of length one), Default: '>1'}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: 'uid_chr'}
}
\value{
Drop offs (a character vector)
}
\description{
get_drop_offs() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get drop offs. The function returns Drop offs (a character vector).
}
\keyword{internal}
