% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fl_meta_from_dv_ls}
\alias{get_fl_meta_from_dv_ls}
\title{Get file meta from dataverse list}
\usage{
get_fl_meta_from_dv_ls(
  ds_ls,
  fl_nm_1L_chr,
  nms_chr = NA_character_,
  type_1L_chr = "description"
)
}
\arguments{
\item{ds_ls}{Dataset (a list)}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{nms_chr}{Names (a character vector), Default: 'NA'}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'description'}
}
\value{
Metadata (an output object of multiple potential types)
}
\description{
get_fl_meta_from_dv_ls() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get file meta from dataverse list. The function returns Metadata (an output object of multiple potential types).
}
\keyword{internal}
