% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fl_nms_of_types}
\alias{get_fl_nms_of_types}
\title{Get file names of types}
\usage{
get_fl_nms_of_types(fl_nms_chr, types_chr)
}
\arguments{
\item{fl_nms_chr}{File names (a character vector)}

\item{types_chr}{Types (a character vector)}
}
\value{
Subset of file names (a character vector)
}
\description{
get_fl_nms_of_types() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get file names of types. The function returns Subset of file names (a character vector).
}
\keyword{internal}
