% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_get_import_type_ls.R
\name{get_import_type_ls}
\alias{get_import_type_ls}
\alias{get_import_type_ls.ready4_all_import_lup}
\alias{get_import_type_ls,ready4_all_import_lup-method}
\title{Get import type list}
\usage{
get_import_type_ls(x, ...)

get_import_type_ls.ready4_all_import_lup(
  x,
  inc_script_lgl = T,
  forced_choice_chr = NA_character_
)

\S4method{get_import_type_ls}{ready4_all_import_lup}(x, inc_script_lgl = T, forced_choice_chr = NA_character_)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{...}{Additional arguments}

\item{inc_script_lgl}{Include script (a logical vector), Default: T}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}
}
\description{
get_import_type_ls() is a Get Import Type List generic that retrieves data about the type of import to be processed.

get_import_type_ls.ready4_all_import_lup() is a Get Import Type List method that retrieves data about the type of import to be processed. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.. The function is called for its side effects and does not return a value.
}
