% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{get_import_type_ls}
\alias{get_import_type_ls}
\title{Get import type list}
\usage{
get_import_type_ls(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
get_import_type_ls() is a Get Import Type List generic that retrieves data about the type of import to be processed.NA
}
\keyword{internal}
