% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_journal_plot_fn}
\alias{get_journal_plot_fn}
\title{Get journal plot function}
\usage{
get_journal_plot_fn(
  what_1L_chr = "barplot",
  pkg_1L_chr = "ggpubr",
  prefix_1L_chr = "gg"
)
}
\arguments{
\item{what_1L_chr}{What (a character vector of length one), Default: 'barplot'}

\item{pkg_1L_chr}{Package (a character vector of length one), Default: 'ggpubr'}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: 'gg'}
}
\value{
Journal plot (an output object of multiple potential types)
}
\description{
get_journal_plot_fn() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get journal plot function. The function returns Journal plot (an output object of multiple potential types).
}
\keyword{internal}
