% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_r3_from_dv_csv}
\alias{get_r3_from_dv_csv}
\title{Get ready4 submodule from dataverse comma separated variables file}
\usage{
get_r3_from_dv_csv(
  file_name_chr,
  data_repo_db_ui_chr,
  data_repo_ui_chr = NA_character_,
  r3_fn = ready4use_imports
)
}
\arguments{
\item{file_name_chr}{File name (a character vector)}

\item{data_repo_db_ui_chr}{Data repository database user interface (a character vector)}

\item{data_repo_ui_chr}{Data repository user interface (a character vector), Default: 'NA'}

\item{r3_fn}{Ready4 submodule (a function), Default: ready4use_imports}
}
\value{
Tibble ready4 submodule (a ready4 submodule extension of tibble)
}
\description{
get_r3_from_dv_csv() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get ready4 submodule from dataverse comma separated variables file. The function returns Tibble ready4 submodule (a ready4 submodule extension of tibble).
}
