% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_raw_dss}
\alias{get_raw_dss}
\title{Get raw datasets}
\usage{
get_raw_dss(paths_chr, names_chr, read_fn = read.csv)
}
\arguments{
\item{paths_chr}{Paths (a character vector)}

\item{names_chr}{Names (a character vector)}

\item{read_fn}{Read (a function), Default: read.csv}
}
\value{
Dataset (a list)
}
\description{
get_raw_dss() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get raw datasets. The function returns Dataset (a list).
}
\keyword{internal}
