% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{get_read_fn}
\alias{get_read_fn}
\title{Get read function}
\usage{
get_read_fn(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
get_read_fn() is a Get Read Function generic that retrieves a read function.
}
