% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_get_read_fn.R
\name{get_read_fn.ready4_dv_import_lup}
\alias{get_read_fn.ready4_dv_import_lup}
\title{Get read function method applied toeadyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse..}
\usage{
\method{get_read_fn}{ready4_dv_import_lup}(x)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.}
}
\description{
get_read_fn.ready4_dv_import_lup() is a Get Read Function method that retrieves a read function. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse..NA
}
\keyword{internal}
