% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_reference_descs}
\alias{get_reference_descs}
\title{Get reference descriptions}
\usage{
get_reference_descs(
  correspondences_ls,
  correspondences_r3 = ready4show::ready4show_correspondences()
)
}
\arguments{
\item{correspondences_ls}{Correspondences (a list)}

\item{correspondences_r3}{Correspondences (a ready4 submodule), Default: ready4show::ready4show_correspondences()}
}
\value{
Reference descriptions (a character vector)
}
\description{
get_reference_descs() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get reference descriptions. The function returns Reference descriptions (a character vector).
}
\keyword{internal}
