% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_styles}
\alias{get_styles}
\title{Get styles}
\usage{
get_styles(what_1L_chr = c("all", "ggsci", "viridis"), sort_1L_lgl = FALSE)
}
\arguments{
\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "ggsci", "viridis")}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: FALSE}
}
\value{
Styles (a character vector)
}
\description{
get_styles() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get styles. The function returns Styles (a character vector).
}
