% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_valid_path_chr}
\alias{get_valid_path_chr}
\title{Get valid path}
\usage{
get_valid_path_chr(x)
}
\arguments{
\item{x}{An object}
}
\value{
Valid path (a character vector)
}
\description{
get_valid_path_chr() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get valid path character vector. Function argument x specifies the where to look for the required object. The function returns Valid path (a character vector).
}
\keyword{internal}
