% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_vars_with_cdn}
\alias{get_vars_with_cdn}
\title{Get variables with condition}
\usage{
get_vars_with_cdn(data_tb, cdn_fn)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{cdn_fn}{Condition (a function)}
}
\value{
Variables (a character vector)
}
\description{
get_vars_with_cdn() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get variables with condition. The function returns Variables (a character vector).
}
\keyword{internal}
