% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_import_chr_vec.R
\name{import_chr_vec}
\alias{import_chr_vec}
\alias{import_chr_vec-ready4_local_proc}
\title{import_chr_vec}
\usage{
import_chr_vec(x)

\S4method{import_chr_vec}{ready4_local_proc}(x)
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
S4 Generic function to get the value of the slot import_chr_vec

Get the value of the slot import_chr_vec for S4 objects of class ready4_local_proc
}
