% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
import_data() is an Import Data generic that imports data from saved files and loads them into memory as R objects.
}
