% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_import_this_ls.R
\name{import_this_ls}
\alias{import_this_ls}
\alias{import_this_ls-ready4_local_proc}
\title{import_this_ls}
\usage{
import_this_ls(x)

\S4method{import_this_ls}{ready4_local_proc}(x)
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
S4 Generic function to get the value of the slot import_this_ls

Get the value of the slot import_this_ls for S4 objects of class ready4_local_proc
}
