% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ingest.R
\name{ingest-Ready4useRepos}
\alias{ingest-Ready4useRepos}
\alias{ingest,Ready4useRepos-method}
\title{ingest}
\usage{
\S4method{ingest}{Ready4useRepos}(x, idcs_int = NA_integer_, key_1L_chr = NULL)
}
\description{
ingest method applied to Ready4useRepos
}
