% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dictionary.R
\name{is_ready4_dictionary}
\alias{is_ready4_dictionary}
\title{Is ready4 s3 class defining a data dictionary tibble.}
\usage{
is_ready4_dictionary(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 s3 class defining a data dictionary tibble.
}
\description{
Check whether an object is a valid instance of the ready4 s3 class defining a data dictionary tibble.
}
\details{
ready4 s3 class defining a data dictionary tibble.
}
