% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dv_import_lup.R
\name{is_ready4_dv_import_lup}
\alias{is_ready4_dv_import_lup}
\title{Is ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.}
\usage{
is_ready4_dv_import_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\details{
ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
