% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_struc_mape.R
\name{is_ready4_par_struc_mape}
\alias{is_ready4_par_struc_mape}
\title{Is Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
is_ready4_par_struc_mape(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Check whether an object is a valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
