% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_all_import_lup.R
\name{is_ready4use_all_import_lup}
\alias{is_ready4use_all_import_lup}
\title{Is ready4use all import lookup table ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}
\usage{
is_ready4use_all_import_lup(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\details{
ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
