% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_distributions.R
\name{is_ready4use_distributions}
\alias{is_ready4use_distributions}
\title{Is ready4use package distributions ready4 S3 class for list object that summarises the parameters of each distribution}
\usage{
is_ready4use_distributions(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for list object that summarises the parameters of each distribution
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for list object that summarises the parameters of each distribution
}
\details{
ready4 S3 class for list object that summarises the parameters of each distribution
}
