% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_imports.R
\name{is_ready4use_imports}
\alias{is_ready4use_imports}
\title{is ready4use package imports ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}
\usage{
is_ready4use_imports(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\details{
Tibble object lookup table of sources of raw (un-processed) data to import.
}
