% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_manifest.R
\name{is_ready4use_manifest}
\alias{is_ready4use_manifest}
\title{Is ready4use package manifest ready4 s3 class defining a manifest of data required to create an R package.}
\usage{
is_ready4use_manifest(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 s3 class defining a manifest of data required to create an R package.
}
\description{
Check whether an object is a valid instance of the ready4 s3 class defining a manifest of data required to create an R package.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
