% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_mapes.R
\name{is_ready4use_mapes}
\alias{is_ready4use_mapes}
\title{is ready4use package mean absolute prediction errors ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
is_ready4use_mapes(x)
}
\arguments{
\item{x}{An object of any type}
}
\value{
A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Check whether an object is a valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
