% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make.ready4use_all_import_lup}
\alias{make.ready4use_all_import_lup}
\title{Make method applied to ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import..}
\usage{
make.ready4use_all_import_lup(
  x,
  forced_choice_chr = NA_character_,
  script_args_ls = NULL
)
}
\arguments{
\item{x}{An object}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{script_args_ls}{Script arguments (a list), Default: NULL}
}
\description{
make.ready4use_all_import_lup() is a Make method that creates a new R object. This method is implemented for the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.
}
