% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make.ready4use_sp_import_lup}
\alias{make.ready4use_sp_import_lup}
\title{Make.ready4use sp import lookup table}
\usage{
make.ready4use_sp_import_lup(x)
}
\arguments{
\item{x}{An object}
}
\description{
make.ready4use_sp_import_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make.ready4use sp import lookup table. The function is called for its side effects and does not return a value.
}
