% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_correspondences}
\alias{make_correspondences}
\title{Make correspondences}
\usage{
make_correspondences(
  dyad_ls,
  append_1L_lgl = T,
  correspondences_ls = NULL,
  names_1L_lgl = F,
  reference_1L_int = 2L
)
}
\arguments{
\item{dyad_ls}{Dyad (a list)}

\item{append_1L_lgl}{Append (a logical vector of length one), Default: T}

\item{correspondences_ls}{Correspondences (a list), Default: NULL}

\item{names_1L_lgl}{Names (a logical vector of length one), Default: F}

\item{reference_1L_int}{Reference (an integer vector of length one), Default: 2}
}
\value{
Correspondences (a list)
}
\description{
make_correspondences() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make correspondences. The function returns Correspondences (a list).
}
\keyword{internal}
