% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{make_dv_import_lup}
\alias{make_dv_import_lup}
\title{Make dataverse import lookup table}
\usage{
make_dv_import_lup(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
make_dv_import_lup() is a Make Dataverse Import Lookup Table generic that makes a Dataverse import lookup tableNA
}
\keyword{internal}
