% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_make_import_xx.R
\name{make_import_xx}
\alias{make_import_xx}
\alias{make_import_xx.ready4_all_import_lup}
\alias{make_import_xx,ready4_all_import_lup-method}
\title{Make import output object of multiple potential types}
\usage{
make_import_xx(x, ...)

make_import_xx.ready4_all_import_lup(
  x,
  forced_choice_chr = NA_character_,
  script_args_ls = NULL
)

\S4method{make_import_xx}{ready4_all_import_lup}(x, forced_choice_chr = NA_character_, script_args_ls = NULL)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{...}{Additional arguments}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{script_args_ls}{Script arguments (a list), Default: NULL}
}
\description{
make_import_xx() is a Make Import Output Object of Multiple Potential Types generic that makes an output object of multiple potential classes.

make_import_xx.ready4_all_import_lup() is a Make Import Output Object of Multiple Potential Types method that makes an output object of multiple potential classes. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.. The function is called for its side effects and does not return a value.
}
