% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{make_import_xx}
\alias{make_import_xx}
\title{Make import output object of multiple potential types}
\usage{
make_import_xx(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
make_import_xx() is a Make Import Output Object of Multiple Potential Types generic that makes an output object of multiple potential classes.NA
}
\keyword{internal}
