% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_make_import_xx.R
\name{make_import_xx.ready4_all_import_lup}
\alias{make_import_xx.ready4_all_import_lup}
\title{Make import output object of multiple potential types method applied toeadyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import..}
\usage{
\method{make_import_xx}{ready4_all_import_lup}(x, forced_choice_chr = NA_character_, script_args_ls = NULL)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{script_args_ls}{Script arguments (a list), Default: NULL}
}
\description{
make_import_xx.ready4_all_import_lup() is a Make Import Output Object of Multiple Potential Types method that makes an output object of multiple potential classes. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.. The function is called for its side effects and does not return a value.
}
