% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_imputed_distinct_cases}
\alias{make_imputed_distinct_cases}
\title{Make imputed distinct cases}
\usage{
make_imputed_distinct_cases(
  data_tb,
  method_1L_chr = c("first", "sample"),
  uid_1L_chr = "UID_chr"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{method_1L_chr}{Method (a character vector of length one), Default: c("first", "sample")}

\item{uid_1L_chr}{Unique identifier (a character vector of length one), Default: 'UID_chr'}
}
\value{
Distinct (a tibble)
}
\description{
make_imputed_distinct_cases() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make imputed distinct cases. The function returns Distinct (a tibble).
}
