% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_keep_lgl}
\alias{make_keep_lgl}
\title{Make keep logical vector}
\usage{
make_keep_lgl(ds_tb, filter_fn = is.na, summary_fn = any, var_nms_chr)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{filter_fn}{Filter (a function), Default: is.na}

\item{summary_fn}{Summary (a function), Default: any}

\item{var_nms_chr}{Variable names (a character vector)}
}
\value{
Keep (a logical vector)
}
\description{
make_keep_lgl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make keep logical vector. The function returns Keep (a logical vector).
}
\keyword{internal}
