% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_missing_report}
\alias{make_missing_report}
\title{Make missing report}
\usage{
make_missing_report(ds_xx)
}
\arguments{
\item{ds_xx}{Dataset (an output object of multiple potential types)}
}
\value{
Missing report (an output object of multiple potential types)
}
\description{
make_missing_report() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make missing report. The function returns Missing report (an output object of multiple potential types).
}
