% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_dictionary.R
\name{make_new_ready4use_dictionary}
\alias{make_new_ready4use_dictionary}
\title{Make new ready4use package dictionary ready4 s3 class defining a data dictionary tibble.}
\usage{
make_new_ready4use_dictionary(x)
}
\arguments{
\item{x}{A prototype for the ready4 s3 class defining a data dictionary tibble.}
}
\value{
An unvalidated instance of the ready4 s3 class defining a data dictionary tibble.
}
\description{
Create a new unvalidated instance of the ready4 s3 class defining a data dictionary tibble.
}
\details{
ready4 s3 class defining a data dictionary tibble.
}
