% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_param_struc_mape.R
\name{make_new_ready4use_param_struc_mape}
\alias{make_new_ready4use_param_struc_mape}
\title{Make new ready4use package parameter struc mean absolute prediction error ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
make_new_ready4use_param_struc_mape(x)
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
}
\value{
An unvalidated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new unvalidated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
