% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_period_correspondences}
\alias{make_period_correspondences}
\title{Make period correspondences}
\usage{
make_period_correspondences(
  descriptions_chr,
  integers_1L_lgl = TRUE,
  name_1L_chr = "period",
  plural_chr = "s",
  range_int = 1L:12L,
  reference_1L_int = integer(0),
  spaced_1L_lgl = TRUE,
  type_1L_chr = "descriptions",
  units_chr = c("minute", "hour", "week", "month", "year")
)
}
\arguments{
\item{descriptions_chr}{Descriptions (a character vector)}

\item{integers_1L_lgl}{Integers (a logical vector of length one), Default: TRUE}

\item{name_1L_chr}{Name (a character vector of length one), Default: 'period'}

\item{plural_chr}{Plural (a character vector), Default: 's'}

\item{range_int}{Range (an integer vector), Default: 1L:12L}

\item{reference_1L_int}{Reference (an integer vector of length one), Default: integer(0)}

\item{spaced_1L_lgl}{Spaced (a logical vector of length one), Default: TRUE}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'descriptions'}

\item{units_chr}{Units (a character vector), Default: c("minute", "hour", "week", "month", "year")}
}
\value{
Correspondences (an output object of multiple potential types)
}
\description{
make_period_correspondences() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make period correspondences. The function returns Correspondences (an output object of multiple potential types).
}
\keyword{internal}
