% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_all_import_lup.R
\name{make_prototype_ready4_all_import_lup}
\alias{make_prototype_ready4_all_import_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}
\usage{
make_prototype_ready4_all_import_lup()
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
