% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dist.R
\name{make_prototype_ready4_dist}
\alias{make_prototype_ready4_dist}
\title{Make prototype Readyforwhatsnext S3 class for list object that summarises the parameters of each distribution}
\usage{
make_prototype_ready4_dist()
}
\value{
A prototype for Readyforwhatsnext S3 class for list object that summarises the parameters of each distribution
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for list object that summarises the parameters of each distribution
}
\details{
Readyforwhatsnext S3 class for list object that summarises the parameters of each distribution
}
