% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dv_import_lup.R
\name{make_prototype_ready4_dv_import_lup}
\alias{make_prototype_ready4_dv_import_lup}
\title{Make prototype Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.}
\usage{
make_prototype_ready4_dv_import_lup()
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.
}
