% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_struc_mape.R
\name{make_prototype_ready4_par_struc_mape}
\alias{make_prototype_ready4_par_struc_mape}
\title{Make prototype Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
make_prototype_ready4_par_struc_mape()
}
\value{
A prototype for Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new prototype for the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
