% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dv_import_lup.R
\name{make_pt_ready4_dv_import_lup}
\alias{make_pt_ready4_dv_import_lup}
\title{Make prototype ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.}
\usage{
make_pt_ready4_dv_import_lup(
  file_type = character(0),
  file_name = character(0),
  data_repo = character(0),
  data_repo_ui = character(0),
  data_repo_db_ui = character(0),
  data_repo_file_ext = character(0),
  data_repo_save_type = character(0)
)
}
\arguments{
\item{file_type}{PARAM_DESCRIPTION, Default: character(0)}

\item{file_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_ui}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_db_ui}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_file_ext}{PARAM_DESCRIPTION, Default: character(0)}

\item{data_repo_save_type}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for
}
\description{
Create a new prototype for the
}
\details{
ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
