% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_dataverses.R
\name{make_pt_ready4use_dataverses}
\alias{make_pt_ready4use_dataverses}
\title{Make prototype ready4use package dataverses ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.}
\usage{
make_pt_ready4use_dataverses(
  file_type_chr = character(0),
  file_name_chr = character(0),
  data_repo_chr = character(0),
  data_repo_ui_chr = character(0),
  data_repo_db_ui_chr = character(0),
  data_repo_file_ext_chr = character(0),
  data_repo_save_type_chr = character(0)
)
}
\arguments{
\item{file_type_chr}{File type (a character vector), Default: character(0)}

\item{file_name_chr}{File name (a character vector), Default: character(0)}

\item{data_repo_chr}{Data repository (a character vector), Default: character(0)}

\item{data_repo_ui_chr}{Data repository user interface (a character vector), Default: character(0)}

\item{data_repo_db_ui_chr}{Data repository database user interface (a character vector), Default: character(0)}

\item{data_repo_file_ext_chr}{Data repository file extension (a character vector), Default: character(0)}

\item{data_repo_save_type_chr}{Data repository save type (a character vector), Default: character(0)}
}
\value{
A prototype for ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\description{
Create a new prototype for the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\details{
ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
