% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_r3_from_csv_tb}
\alias{make_r3_from_csv_tb}
\title{Make readyforwhatsnext S3 from csv}
\usage{
make_r3_from_csv_tb(csv_tb, r3_fn)
}
\arguments{
\item{csv_tb}{Csv (a tibble)}

\item{r3_fn}{Readyforwhatsnext S3 (a function)}
}
\value{
Tibble readyforwhatsnext S3 (a readyforwhatsnext S3 extension of tibble)
}
\description{
make_r3_from_csv_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make a readyforwhatsnext S3 from csv.The function returns a tibble readyforwhatsnext s3 (a readyforwhatsnext s3 extension of tibble).
}
\keyword{internal}
