% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_significance_df}
\alias{make_significance_df}
\title{Make significance dataframe}
\usage{
make_significance_df(data_tb, by_1L_chr, vars_chr, sort_1L_lgl = T)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{by_1L_chr}{By (a character vector of length one)}

\item{vars_chr}{Variables (a character vector)}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: T}
}
\value{
Data (a data.frame)
}
\description{
make_significance_df() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make significance dataframe. The function returns Data (a data.frame).
}
\keyword{internal}
