% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_temporal_lup}
\alias{make_temporal_lup}
\title{Make temporal lookup table}
\usage{
make_temporal_lup(dyad_ls, recode_ls, spaced_1L_lgl = TRUE)
}
\arguments{
\item{dyad_ls}{Dyad (a list)}

\item{recode_ls}{Recode (a list)}

\item{spaced_1L_lgl}{Spaced (a logical vector of length one), Default: TRUE}
}
\value{
Temporal (a lookup table)
}
\description{
make_temporal_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make temporal lookup table. The function returns Temporal (a lookup table).
}
\keyword{internal}
