% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_merge_itms_chr.R
\name{merge_itms_chr}
\alias{merge_itms_chr}
\alias{merge_itms_chr-Ready4useFiles}
\alias{merge_itms_chr,Ready4useFiles-method}
\title{merge_itms_chr}
\usage{
merge_itms_chr(x)

\S4method{merge_itms_chr}{Ready4useFiles}(x)
}
\arguments{
\item{x}{An object of class Ready4useFiles}
}
\description{
S4 Generic function to get the value of the slot merge_itms_chr

Get the value of the slot merge_itms_chr for S4 objects of class Ready4useFiles
}
