% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_merge_itms_chr.R
\name{merge_itms_chr<-}
\alias{merge_itms_chr<-}
\alias{merge_itms_chr<--Ready4useFiles}
\alias{merge_itms_chr<-,Ready4useFiles-method}
\title{merge_itms_chr<-}
\usage{
merge_itms_chr(x) <- value

\S4method{merge_itms_chr}{Ready4useFiles}(x) <- value
}
\arguments{
\item{x}{An object of class Ready4useFiles}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot merge_itms_chr

Set the value of the slot merge_itms_chr for S4 objects of class Ready4useFiles
}
