% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_merge_with_chr_vec.R
\name{merge_with_chr_vec}
\alias{merge_with_chr_vec}
\title{merge_with_chr_vec}
\usage{
merge_with_chr_vec(x)
}
\arguments{
\item{x}{An object}
}
\description{
S4 Generic function to get the value of the slot merge_with_chr_vec
}
