% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_merge_with_chr_vec.R
\name{merge_with_chr_vec<-}
\alias{merge_with_chr_vec<-}
\alias{merge_with_chr_vec<--ready4use_local}
\alias{merge_with_chr_vec<-,ready4use_local-method}
\title{merge_with_chr_vec<-}
\usage{
merge_with_chr_vec(x) <- value

\S4method{merge_with_chr_vec}{ready4use_local}(x) <- value
}
\arguments{
\item{x}{An object of class ready4use_local}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot merge_with_chr_vec

Set the value of the slot merge_with_chr_vec for S4 objects of class ready4use_local
}
