% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_metamorphose.R
\name{metamorphose.ready4use_manifest}
\alias{metamorphose.ready4use_manifest}
\alias{metamorphose,ready4use_manifest-method}
\title{Metamorphose - a method that transforms an instance of a class into an object with different structural properties.}
\usage{
metamorphose.ready4use_manifest(x)

\S4method{metamorphose}{ready4use_manifest}(x)
}
\arguments{
\item{x}{An instance of ready4 s3 class defining a manifest of data required to create an R package.}
}
\value{
Manifest (a ready4 S3)
}
\description{
metamorphose.ready4use_manifest() is a Metamorphose method that transforms an instance of a class into an object with different structural properties. This method is implemented for the ready4 s3 class defining a manifest of data required to create an R package. The function returns Manifest (a ready4 S3).
}
