% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_overwrite_lgl.R
\name{overwrite_lgl<-}
\alias{overwrite_lgl<-}
\alias{overwrite_lgl<--ready4_local}
\title{overwrite_lgl<-}
\usage{
overwrite_lgl(x) <- value

\S4method{overwrite_lgl}{ready4_local}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_local}
}
\description{
S4 Generic function to set the value of the slot overwrite_lgl

Set the value of the slot overwrite_lgl for S4 objects of class ready4_local
}
