% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_path_to_starter_sf_chr.R
\name{path_to_starter_sf_chr<-}
\alias{path_to_starter_sf_chr<-}
\alias{path_to_starter_sf_chr<--ready4_local_proc}
\title{path_to_starter_sf_chr<-}
\usage{
path_to_starter_sf_chr(x) <- value

\S4method{path_to_starter_sf_chr}{ready4_local_proc}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
S4 Generic function to set the value of the slot path_to_starter_sf_chr

Set the value of the slot path_to_starter_sf_chr for S4 objects of class ready4_local_proc
}
