% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_pckg_chr.R
\name{pckg_chr<-}
\alias{pckg_chr<-}
\alias{pckg_chr<--ready4_local}
\title{pckg_chr<-}
\usage{
pckg_chr(x) <- value

\S4method{pckg_chr}{ready4_local}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_local}
}
\description{
S4 Generic function to set the value of the slot pckg_chr

Set the value of the slot pckg_chr for S4 objects of class ready4_local
}
