% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_pckg_chr.R
\name{pckg_chr}
\alias{pckg_chr}
\alias{pckg_chr-ready4_local}
\title{pckg_chr}
\usage{
pckg_chr(x)

\S4method{pckg_chr}{ready4_local}(x)
}
\arguments{
\item{x}{An object of class ready4_local}
}
\description{
S4 Generic function to get the value of the slot pckg_chr

Get the value of the slot pckg_chr for S4 objects of class ready4_local
}
