% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot.R
\name{plot_for_journal}
\alias{plot_for_journal}
\title{Plot for journal}
\usage{
plot_for_journal(
  data_tb,
  as_percent_1L_lgl = FALSE,
  by_1L_chr = character(0),
  colours_chr = c("#de2d26", "#fc9272"),
  drop_legend_1L_lgl = FALSE,
  drop_missing_1L_lgl = FALSE,
  drop_ticks_1L_lgl = FALSE,
  fill_single_1L_lgl = FALSE,
  flip_1L_lgl = F,
  label_fill_1L_chr = character(0),
  line_1L_chr = "black",
  position_xx = NULL,
  recode_lup_r3 = ready4show::ready4show_correspondences(),
  significance_1L_lgl = F,
  significance_args_ls = list(),
  style_1L_chr = get_styles(),
  title_1L_chr = character(0),
  type_1L_chr = c("ggsci", "manual", "viridis"),
  x_1L_chr = character(0),
  x_label_1L_chr = character(0),
  y_1L_chr = character(0),
  y_label_1L_chr = character(0),
  what_1L_chr = get_journal_plot_fn("names"),
  ...
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{as_percent_1L_lgl}{As percent (a logical vector of length one), Default: FALSE}

\item{by_1L_chr}{By (a character vector of length one), Default: character(0)}

\item{colours_chr}{Colours (a character vector), Default: c("#de2d26", "#fc9272")}

\item{drop_legend_1L_lgl}{Drop legend (a logical vector of length one), Default: FALSE}

\item{drop_missing_1L_lgl}{Drop missing (a logical vector of length one), Default: FALSE}

\item{drop_ticks_1L_lgl}{Drop ticks (a logical vector of length one), Default: FALSE}

\item{fill_single_1L_lgl}{Fill single (a logical vector of length one), Default: FALSE}

\item{flip_1L_lgl}{Flip (a logical vector of length one), Default: F}

\item{label_fill_1L_chr}{Label fill (a character vector of length one), Default: character(0)}

\item{line_1L_chr}{Line (a character vector of length one), Default: 'black'}

\item{position_xx}{Position (an output object of multiple potential types), Default: NULL}

\item{recode_lup_r3}{Recode (a ready4 submodule extension of lookup table), Default: ready4show::ready4show_correspondences()}

\item{significance_1L_lgl}{Significance (a logical vector of length one), Default: F}

\item{significance_args_ls}{Significance arguments (a list), Default: list()}

\item{style_1L_chr}{Style (a character vector of length one), Default: get_styles()}

\item{title_1L_chr}{Title (a character vector of length one), Default: character(0)}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("ggsci", "manual", "viridis")}

\item{x_1L_chr}{X (a character vector of length one), Default: character(0)}

\item{x_label_1L_chr}{X label (a character vector of length one), Default: character(0)}

\item{y_1L_chr}{Y (a character vector of length one), Default: character(0)}

\item{y_label_1L_chr}{Y label (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: get_journal_plot_fn("names")}

\item{...}{Additional arguments}
}
\value{
Plot (a plot)
}
\description{
plot_for_journal() is a Plot function that plots data. Specifically, this function implements an algorithm to plot for journal. The function returns Plot (a plot).
}
\keyword{internal}
