% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_significance}
\alias{print_significance}
\title{Print significance}
\usage{
print_significance(
  data_tb,
  by_1L_chr,
  vars_chr,
  caption_1L_chr = character(0),
  output_type_1L_chr = "HTML",
  sort_1L_lgl = T
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{by_1L_chr}{By (a character vector of length one)}

\item{vars_chr}{Variables (a character vector)}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: character(0)}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{sort_1L_lgl}{Sort (a logical vector of length one), Default: T}
}
\value{
Data (an output object of multiple potential types)
}
\description{
print_significance() is a Print function that prints output to console. Specifically, this function implements an algorithm to print significance. The function returns Data (an output object of multiple potential types).
}
\keyword{internal}
