% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_proc_data_dir_chr.R
\name{proc_data_dir_chr<-}
\alias{proc_data_dir_chr<-}
\alias{proc_data_dir_chr<--ready4_local_proc}
\title{proc_data_dir_chr<-}
\usage{
proc_data_dir_chr(x) <- value

\S4method{proc_data_dir_chr}{ready4_local_proc}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
S4 Generic function to set the value of the slot proc_data_dir_chr

Set the value of the slot proc_data_dir_chr for S4 objects of class ready4_local_proc
}
