% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_proc_data_dir_chr.R
\name{proc_data_dir_chr}
\alias{proc_data_dir_chr}
\alias{proc_data_dir_chr-ready4_local_proc}
\title{proc_data_dir_chr}
\usage{
proc_data_dir_chr(x)

\S4method{proc_data_dir_chr}{ready4_local_proc}(x)
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
S4 Generic function to get the value of the slot proc_data_dir_chr

Get the value of the slot proc_data_dir_chr for S4 objects of class ready4_local_proc
}
