% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure.ready4use_dataverses}
\alias{procure.ready4use_dataverses}
\alias{procure,ready4use_dataverses-method}
\alias{procure.ready4use_imports}
\alias{procure,ready4use_imports-method}
\title{Procure method applied to ready4 S3 class for tibble object lookup table of files to be imported from a dataverse..}
\usage{
procure.ready4use_dataverses(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

\S4method{procure}{ready4use_dataverses}(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

procure.ready4use_imports(
  x,
  inc_script_lgl = T,
  forced_choice_chr = NA_character_
)

\S4method{procure}{ready4use_imports}(x, inc_script_lgl = T, forced_choice_chr = NA_character_)
}
\arguments{
\item{x}{An instance of ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{save_dir_path_1L_chr}{Save directory path (a character vector of length one), Default: ''}

\item{unlink_1L_lgl}{Unlink (a logical vector of length one), Default: T}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{inc_script_lgl}{Include script (a logical vector), Default: T}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}
}
\description{
procure.ready4use_dataverses() is a Procure method that searches and retrieves requested data from a specified source. This method is implemented for the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse. The function is called for its side effects and does not return a value.

procure.ready4use_imports() is a Procure method that searches and retrieves requested data from a specified source. This method is implemented for the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.
}
