% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_raw_data_dir_chr.R
\name{raw_data_dir_chr}
\alias{raw_data_dir_chr}
\alias{raw_data_dir_chr-ready4_local}
\title{raw_data_dir_chr}
\usage{
raw_data_dir_chr(x)

\S4method{raw_data_dir_chr}{ready4_local}(x)
}
\arguments{
\item{x}{An object of class ready4_local}
}
\description{
S4 Generic function to get the value of the slot raw_data_dir_chr

Get the value of the slot raw_data_dir_chr for S4 objects of class ready4_local
}
