% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_all_import_lup.R
\name{ready4_all_import_lup}
\alias{ready4_all_import_lup}
\title{Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}
\usage{
ready4_all_import_lup(x = make_prototype_ready4_all_import_lup())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import., Default: make_prototype_ready4_all_import_lup()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\details{
Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
