% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dv_import_lup.R
\name{ready4_dv_import_lup}
\alias{ready4_dv_import_lup}
\title{ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.}
\usage{
ready4_dv_import_lup(x = make_pt_ready4_dv_import_lup())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse., Default: make_pt_ready4_dv_import_lup()}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\details{
ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
