% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_local_raw.R
\docType{class}
\name{ready4_local_raw}
\alias{ready4_local_raw}
\title{ready4_local_raw}
\usage{
ready4_local_raw(merge_with_chr_vec = NA_character_, save_type = "raw")
}
\arguments{
\item{merge_with_chr_vec}{character, Default: 'NA'}

\item{save_type}{character, Default: 'raw'}
}
\value{
An S4 object of the
}
\description{
An S4 class to represent Readyforwhatsnext S4 class defining data to be saved in local directory in a raw (unprocessed) format.

Create a new S4 object of the class:ready4_local_raw
}
\section{Slots}{

\describe{
\item{\code{save_type}}{character}

\item{\code{merge_with_chr_vec}}{character}
}}

