% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_struc_mape.R
\name{ready4_par_struc_mape}
\alias{ready4_par_struc_mape}
\title{Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
ready4_par_struc_mape(x = make_prototype_ready4_par_struc_mape())
}
\arguments{
\item{x}{A prototype for the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors., Default: make_prototype_ready4_par_struc_mape()}
}
\value{
A validated instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new valid instance of the Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
Readyforwhatsnext S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
