% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_script_data.R
\docType{class}
\name{ready4_script_data}
\alias{ready4_script_data}
\title{ready4_script_data}
\usage{
ready4_script_data(
  save_type = NA_character_,
  proc_data_dir_chr = NA_character_,
  import_chr_vec = NA_character_,
  path_to_starter_sf_chr = NA_character_,
  import_this_ls = list(list()),
  merge_with_chr_vec = NA_character_,
  raw_data_dir_chr = NA_character_,
  pckg_chr = NA_character_,
  overwrite_lgl = NA,
  save_lgl = NA,
  crs_nbr_vec = NA_real_
)
}
\arguments{
\item{save_type}{character, Default: 'NA'}

\item{proc_data_dir_chr}{character, Default: 'NA'}

\item{import_chr_vec}{character, Default: 'NA'}

\item{path_to_starter_sf_chr}{character, Default: 'NA'}

\item{import_this_ls}{list, Default: list(list())}

\item{merge_with_chr_vec}{character, Default: 'NA'}

\item{raw_data_dir_chr}{character, Default: 'NA'}

\item{pckg_chr}{character, Default: 'NA'}

\item{overwrite_lgl}{logical, Default: NA}

\item{save_lgl}{logical, Default: NA}

\item{crs_nbr_vec}{numeric, Default: NA}
}
\value{
An S4 object of the
}
\description{
An S4 class to represent Readyforwhatsnext S4 class containint data to be passed to a function that constructs a spatial object from a lookup table.

Create a new S4 object of the class:ready4_script_data
}
\section{Slots}{

\describe{
\item{\code{crs_nbr_vec}}{numeric}

\item{\code{save_type}}{character}

\item{\code{proc_data_dir_chr}}{character}

\item{\code{import_chr_vec}}{character}

\item{\code{path_to_starter_sf_chr}}{character}

\item{\code{import_this_ls}}{list}

\item{\code{merge_with_chr_vec}}{character}

\item{\code{raw_data_dir_chr}}{character}

\item{\code{pckg_chr}}{character}

\item{\code{overwrite_lgl}}{logical}

\item{\code{save_lgl}}{logical}
}}

