% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4use.R, R/ready4use-package.R
\docType{package}
\name{ready4use-package}
\alias{ready4use-package}
\title{ready4use: Ingest, Label, Visualise and Share Health Economic Model Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ready4use provides a set of tools for managing data for models developed with the ready4 framework. For detailed documentation about the framework and how to use it visit \url{https://www.ready4-dev.com/}. For a background to the methodological issues that the framework is attempting to help solve, see Hamilton et al. (2024) \doi{10.1007/s40273-024-01378-8}. This development version of the ready4use package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ready4use provides a set of tools for managing data for models developed with the ready4 framework. For detailed documentation about the framework and how to use it visit \url{https://www.ready4-dev.com/}. For a background to the methodological issues that the framework is attempting to help solve, see Hamilton et al. (2024) \doi{10.1007/s40273-024-01378-8}. This development version of the ready4use package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4use/}
  \item \url{https://github.com/ready4-dev/ready4use}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4use/}
  \item \url{https://github.com/ready4-dev/ready4use}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Glen Wiesner (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item Australian Government Research Training Program [funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
