% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_all_import_lup.R
\name{ready4use_all_import_lup}
\alias{ready4use_all_import_lup}
\title{ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}
\usage{
ready4use_all_import_lup(x = make_pt_ready4use_all_import_lup())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import., Default: make_pt_ready4use_all_import_lup()}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
\details{
ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.
}
