% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_dataverses.R
\name{ready4use_dataverses}
\alias{ready4use_dataverses}
\title{ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.}
\usage{
ready4use_dataverses(x = make_pt_ready4use_dataverses())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse., Default: make_pt_ready4use_dataverses()}
}
\value{
A validated instance of the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse.
}
\details{
Tibble object lookup table of files to be imported from a dataverse.
}
