% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_dictionary.R
\name{ready4use_dictionary}
\alias{ready4use_dictionary}
\title{ready4 s3 class defining a data dictionary tibble.}
\usage{
ready4use_dictionary(x = make_pt_ready4use_dictionary())
}
\arguments{
\item{x}{A prototype for the ready4 s3 class defining a data dictionary tibble., Default: make_pt_ready4use_dictionary()}
}
\value{
A validated instance of the ready4 s3 class defining a data dictionary tibble.
}
\description{
Create a new valid instance of the ready4 s3 class defining a data dictionary tibble.
}
\details{
ready4 s3 class defining a data dictionary tibble.
}
