% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_manifest.R
\name{ready4use_manifest}
\alias{ready4use_manifest}
\title{ready4 s3 class defining a manifest of data required to create an R package.}
\usage{
ready4use_manifest(x = make_pt_ready4use_manifest())
}
\arguments{
\item{x}{A prototype for the ready4 s3 class defining a manifest of data required to create an R package., Default: make_pt_ready4use_manifest()}
}
\value{
A validated instance of the ready4 s3 class defining a manifest of data required to create an R package.
}
\description{
Create a new valid instance of the ready4 s3 class defining a manifest of data required to create an R package.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
