% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_mapes.R
\name{ready4use_mapes}
\alias{ready4use_mapes}
\title{ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
ready4use_mapes(x = make_pt_ready4use_mapes())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors., Default: make_pt_ready4use_mapes()}
}
\value{
A validated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
Tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
