% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_param_struc_mape.R
\name{ready4use_param_struc_mape}
\alias{ready4use_param_struc_mape}
\title{ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
ready4use_param_struc_mape(x = make_pt_ready4use_param_struc_mape())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors., Default: make_pt_ready4use_param_struc_mape()}
}
\value{
A validated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
