% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4use_script_data.R
\docType{class}
\name{ready4use_script_data}
\alias{ready4use_script_data}
\title{ready4use_script_data}
\description{
An S4 class to represent ready4 S4 class containing data to be passed to a function that constructs a spatial object from a lookup table.
}
\section{Slots}{

\describe{
\item{\code{crs_nbr_dbl}}{numeric}

\item{\code{save_type}}{character}

\item{\code{merge_with_chr_vec}}{character}
}}

