% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useArguments.R
\docType{class}
\name{Ready4useArguments-class}
\alias{Ready4useArguments-class}
\alias{Ready4useArguments}
\title{Ready4useArguments}
\description{
Arguments for a function that constructs a spatial object.
}
\section{Slots}{

\describe{
\item{\code{crs_nbr_dbl}}{Coordinates reference system number (a double vector)}

\item{\code{write_type_1L_chr}}{Write type (a character vector of length one)}

\item{\code{processed_fls_dir_1L_chr}}{Processed files directory (a character vector of length one)}

\item{\code{imports_chr}}{Imports (a character vector)}

\item{\code{path_to_seed_sf_1L_chr}}{Path to seed simple features object (a character vector of length one)}

\item{\code{imports_ls}}{Imports (a list)}

\item{\code{merge_itms_chr}}{Merge items (a character vector)}

\item{\code{raw_fls_dir_1L_chr}}{Raw files directory (a character vector of length one)}

\item{\code{pkg_1L_chr}}{Package (a character vector of length one)}

\item{\code{overwrite_1L_lgl}}{Overwrite (a logical vector of length one)}

\item{\code{write_1L_lgl}}{Write (a logical vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

