% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useArguments.R
\docType{class}
\name{Ready4useArguments}
\alias{Ready4useArguments}
\title{Ready4useArguments}
\description{
An S4 class to represent ready4 S4 class containing data to be passed to a function that constructs a spatial object from a lookup table.
}
\section{Slots}{

\describe{
\item{\code{crs_nbr_dbl}}{numeric}

\item{\code{write_type_1L_chr}}{character}

\item{\code{merge_itms_chr}}{character}
}}

