% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useDyad.R
\docType{class}
\name{Ready4useDyad-class}
\alias{Ready4useDyad-class}
\alias{Ready4useDyad}
\title{Ready4useDyad}
\description{
A dataset and data dictionary pair.
}
\section{Slots}{

\describe{
\item{\code{ds_tb}}{Dataset (a tibble)}

\item{\code{dictionary_r3}}{Dictionary (a ready4 submodule)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

