% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useFiles.R
\docType{class}
\name{Ready4useFiles}
\alias{Ready4useFiles}
\title{Ready4useFiles}
\description{
An S4 class to represent ready4 S4 class defining data to be saved in local directory.
}
\section{Slots}{

\describe{
\item{\code{merge_itms_chr}}{character}
}}

