% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useIngest.R
\docType{class}
\name{Ready4useIngest-class}
\alias{Ready4useIngest-class}
\alias{Ready4useIngest}
\title{Ready4useIngest}
\description{
Ingested data and descriptive metadata.
}
\section{Slots}{

\describe{
\item{\code{objects_ls}}{Objects (a list)}

\item{\code{names_chr}}{Names (a character vector)}

\item{\code{descriptions_chr}}{Descriptions (a character vector)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

