% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4usePointer.R
\docType{class}
\name{Ready4usePointer-class}
\alias{Ready4usePointer-class}
\alias{Ready4usePointer}
\title{Ready4usePointer}
\description{
Metadata on local and remote data storage locations.
}
\section{Slots}{

\describe{
\item{\code{a_Ready4usePaths}}{(an instance of the Ready4Module class)}

\item{\code{b_Ready4useRepos}}{(an instance of the Ready4useRepos class)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

