% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useProcessed.R
\docType{class}
\name{Ready4useProcessed-class}
\alias{Ready4useProcessed-class}
\alias{Ready4useProcessed}
\title{Ready4useProcessed}
\description{
Metadata for dataset(s) to be saved in local directory in a processed (R) format.
}
\section{Slots}{

\describe{
\item{\code{write_type_1L_chr}}{Write type (a character vector of length one)}

\item{\code{processed_fls_dir_1L_chr}}{Processed files directory (a character vector of length one)}

\item{\code{imports_chr}}{Imports (a character vector)}

\item{\code{path_to_seed_sf_1L_chr}}{Path to seed simple features object (a character vector of length one)}

\item{\code{imports_ls}}{Imports (a list)}

\item{\code{merge_itms_chr}}{Merge items (a character vector)}

\item{\code{raw_fls_dir_1L_chr}}{Raw files directory (a character vector of length one)}

\item{\code{pkg_1L_chr}}{Package (a character vector of length one)}

\item{\code{overwrite_1L_lgl}}{Overwrite (a logical vector of length one)}

\item{\code{write_1L_lgl}}{Write (a logical vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

