% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useProcessed.R
\docType{class}
\name{Ready4useProcessed}
\alias{Ready4useProcessed}
\title{Ready4useProcessed}
\description{
An S4 class to represent ready4 S4 class defining data to be saved in local directory in a processed (R) format.
}
\section{Slots}{

\describe{
\item{\code{write_type_1L_chr}}{character}

\item{\code{merge_itms_chr}}{character}
}}

