% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useRaw.R
\docType{class}
\name{Ready4useRaw-class}
\alias{Ready4useRaw-class}
\alias{Ready4useRaw}
\title{Ready4useRaw}
\description{
An S4 class to represent ready4 S4 class defining data to be saved in local directory in a raw (unprocessed) format.
}
\section{Slots}{

\describe{
\item{\code{write_type_1L_chr}}{character}

\item{\code{merge_itms_chr}}{character}

\item{\code{raw_fls_dir_1L_chr}}{character}

\item{\code{pkg_1L_chr}}{character}

\item{\code{overwrite_1L_lgl}}{logical}

\item{\code{write_1L_lgl}}{logical}
}}

