% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useRepos.R
\docType{class}
\name{Ready4useRepos-class}
\alias{Ready4useRepos-class}
\alias{Ready4useRepos}
\title{Ready4useRepos}
\description{
Metadata about online data repositories.
}
\section{Slots}{

\describe{
\item{\code{dataverse_nm_1L_chr}}{Dataverse name (a character vector of length one)}

\item{\code{dv_ds_metadata_ls}}{Dataverse dataset metadata (a list)}

\item{\code{dv_ds_nm_1L_chr}}{Dataverse dataset name (a character vector of length one)}

\item{\code{dv_url_pfx_1L_chr}}{Dataverse url prefix (a character vector of length one)}

\item{\code{rds_objs_nms_chr}}{Rds objects names (a character vector)}

\item{\code{server_1L_chr}}{Server (a character vector of length one)}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

