% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_labels_from_ds}
\alias{remove_labels_from_ds}
\title{Remove labels from dataset}
\usage{
remove_labels_from_ds(ds_tb)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}
}
\value{
Unlabelled dataset (a tibble)
}
\description{
remove_labels_from_ds() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove labels from dataset. Function argument ds_tb specifies the object to be updated. The function returns Unlabelled dataset (a tibble).
}
\keyword{internal}
