% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.ready4use_dictionary}
\alias{renew.ready4use_dictionary}
\alias{renew,ready4use_dictionary-method}
\alias{renew.ready4use_imports}
\alias{renew,ready4use_imports-method}
\title{Renew method applied to ready4 S3 class defining a data dictionary tibble..}
\usage{
renew.ready4use_dictionary(x, new_ready4_dict_r3)

\S4method{renew}{ready4use_dictionary}(x, new_ready4_dict_r3)

renew.ready4use_imports(x, local_to_url_vec_chr, urls_vec_chr)

\S4method{renew}{ready4use_imports}(x, local_to_url_vec_chr, urls_vec_chr)
}
\arguments{
\item{x}{An instance of ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{new_ready4_dict_r3}{New ready4 dictionary (a ready4 S3)}

\item{local_to_url_vec_chr}{Local to url vector (a character vector)}

\item{urls_vec_chr}{Urls vector (a character vector)}
}
\value{
combined_ready4_dictionaries (An object)
}
\description{
renew.ready4use_dictionary() is a Renew method that updates an instance of a class with new values. This method is implemented for the ready4 s3 class defining a data dictionary tibble. The function is called for its side effects and does not return a value.

renew.ready4use_imports() is a Renew method that updates an instance of a class with new values. This method is implemented for the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.
}
