% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_save_lgl.R
\name{save_lgl}
\alias{save_lgl}
\alias{save_lgl-ready4_local}
\title{save_lgl}
\usage{
save_lgl(x)

\S4method{save_lgl}{ready4_local}(x)
}
\arguments{
\item{x}{An object of class ready4_local}
}
\description{
S4 Generic function to get the value of the slot save_lgl

Get the value of the slot save_lgl for S4 objects of class ready4_local
}
