% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{save_raw}
\alias{save_raw}
\title{Save raw}
\usage{
save_raw(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
save_raw() is a Save Raw generic that saves the native version of a file format.NA
}
