% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_save_type.R
\name{save_type-ready4_local_raw}
\alias{save_type-ready4_local_raw}
\alias{save_type,ready4_local_raw-method}
\alias{save_type-ready4_local_proc}
\alias{save_type,ready4_local_proc-method}
\title{save_type}
\usage{
\S4method{save_type}{ready4_local_raw}(x)

\S4method{save_type}{ready4_local_proc}(x)
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
Get the value of the slot save_type for S4 objects of class ready4_local_raw

Get the value of the slot save_type for S4 objects of class ready4_local_proc
}
