% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_save_type.R
\name{save_type<--ready4_local_raw}
\alias{save_type<--ready4_local_raw}
\alias{save_type<--ready4_local_proc}
\title{save_type<-}
\usage{
\S4method{save_type}{ready4_local_raw}(x) <- value

\S4method{save_type}{ready4_local_proc}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_local_proc}
}
\description{
Set the value of the slot save_type for S4 objects of class ready4_local_raw

Set the value of the slot save_type for S4 objects of class ready4_local_proc
}
