% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_share.R
\name{share-Ready4useRecord}
\alias{share-Ready4useRecord}
\alias{share,Ready4useRecord-method}
\title{share}
\usage{
\S4method{share}{Ready4useRecord}(
  x,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  type_1L_chr = "dataverse",
  publish_dv_1L_lgl = F
)
}
\description{
share method applied to Ready4useRecord
}
