% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_csv_col_to_ls_col}
\alias{transform_csv_col_to_ls_col}
\title{Transform comma separated variables file column to list column}
\usage{
transform_csv_col_to_ls_col(csv_col)
}
\arguments{
\item{csv_col}{PARAM_DESCRIPTION}
}
\value{
List column (a list)
}
\description{
transform_csv_col_to_ls_col() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform comma separated variables file column to list column. Function argument csv_col specifies the object to be updated. The function returns List column (a list).
}
