% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_dates}
\alias{transform_dates}
\title{Transform dates}
\usage{
transform_dates(dates_chr, drop_time_1L_lgl = F)
}
\arguments{
\item{dates_chr}{Dates (a character vector)}

\item{drop_time_1L_lgl}{Drop time (a logical vector of length one), Default: F}
}
\value{
Dates (a date vector)
}
\description{
transform_dates() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dates. The function returns Dates (a date vector).
}
\keyword{internal}
