% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_dyad_to_series}
\alias{transform_dyad_to_series}
\title{Transform dyad to series}
\usage{
transform_dyad_to_series(
  X_Ready4useDyad = Ready4useDyad(),
  timepoint_vals_chr = c("baseline", "follow_up"),
  id_var_nm_1L_chr = "uid_chr",
  participation_var_1L_chr = "participation",
  timepoint_var_nm_1L_chr = "timing_fct"
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: Ready4useDyad()}

\item{timepoint_vals_chr}{Timepoint values (a character vector), Default: c("baseline", "follow_up")}

\item{id_var_nm_1L_chr}{Identity variable name (a character vector of length one), Default: 'uid_chr'}

\item{participation_var_1L_chr}{Participation variable (a character vector of length one), Default: 'participation'}

\item{timepoint_var_nm_1L_chr}{Timepoint variable name (a character vector of length one), Default: 'timing_fct'}
}
\value{
Z (A longitudinal dataset and its associated dictionary, descriptive statistics and metadata.)
}
\description{
transform_dyad_to_series() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform dyad to series. The function is called for its side effects and does not return a value.
}
\keyword{internal}
