% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_raw_df}
\alias{transform_raw_df}
\title{Transform raw dataframe}
\usage{
transform_raw_df(data_df, cases_start_at_1L_int = 1L)
}
\arguments{
\item{data_df}{Data (a data.frame)}

\item{cases_start_at_1L_int}{Cases start at (an integer vector of length one), Default: 1}
}
\value{
Dataset (a tibble)
}
\description{
transform_raw_df() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform raw dataframe. The function returns Dataset (a tibble).
}
\keyword{internal}
