% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_character_vars}
\alias{update_character_vars}
\title{Update character variables}
\usage{
update_character_vars(
  ds_tb,
  var_nms_chr,
  as_missing_chr = character(0),
  missing_1L_chr = NA_character_,
  prefix_1L_chr = character(0),
  remove_end_chr = character(0),
  remove_start_chr = character(0),
  replacement_fn_ls = list(end = stringi::stri_replace_last_fixed, start =
    stringi::stri_replace_first_fixed),
  x_ready4show_correspondences = ready4show::ready4show_correspondences()
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{var_nms_chr}{Variable names (a character vector)}

\item{as_missing_chr}{As missing (a character vector), Default: character(0)}

\item{missing_1L_chr}{Missing (a character vector of length one), Default: 'NA'}

\item{prefix_1L_chr}{Prefix (a character vector of length one), Default: character(0)}

\item{remove_end_chr}{Remove end (a character vector), Default: character(0)}

\item{remove_start_chr}{Remove start (a character vector), Default: character(0)}

\item{replacement_fn_ls}{Replacement (a list of functions), Default: list(end = stringi::stri_replace_last_fixed, start = stringi::stri_replace_first_fixed)}

\item{x_ready4show_correspondences}{PARAM_DESCRIPTION, Default: ready4show::ready4show_correspondences()}
}
\value{
Dataset (a tibble)
}
\description{
update_character_vars() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update character variables. The function returns Dataset (a tibble).
}
\keyword{internal}
