% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_data_dict}
\alias{update_data_dict}
\title{Update data dictionary}
\usage{
update_data_dict(
  X_Ready4useDyad = Ready4useDyad(),
  dictionary_lups_ls = list(),
  arrange_by_1L_chr = c("category", "var_ctg_chr", "name", "var_nm_chr", "both",
    "var_ctg_chr, var_nm_chr")
)
}
\arguments{
\item{X_Ready4useDyad}{PARAM_DESCRIPTION, Default: Ready4useDyad()}

\item{dictionary_lups_ls}{Dictionary lookup tables (a list), Default: list()}

\item{arrange_by_1L_chr}{Arrange by (a character vector of length one), Default: c("category", "var_ctg_chr", "name", "var_nm_chr", "both", "var_ctg_chr, var_nm_chr")}
}
\value{
X (A dataset and data dictionary pair.)
}
\description{
update_data_dict() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update data dictionary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
