% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_pairs_ls}
\alias{update_pairs_ls}
\title{Update pairs list}
\usage{
update_pairs_ls(
  pairs_ls,
  append_ls = NULL,
  correspondences_r3 = ready4show::ready4show_correspondences(),
  datestamp_chr = character(0),
  discard_chr = character(0)
)
}
\arguments{
\item{pairs_ls}{Pairs (a list)}

\item{append_ls}{Append (a list), Default: NULL}

\item{correspondences_r3}{Correspondences (a ready4 submodule), Default: ready4show::ready4show_correspondences()}

\item{datestamp_chr}{Datestamp (a character vector), Default: character(0)}

\item{discard_chr}{Discard (a character vector), Default: character(0)}
}
\value{
Pairs (a list)
}
\description{
update_pairs_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update pairs list. The function returns Pairs (a list).
}
\keyword{internal}
