% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_raw_data}
\alias{update_raw_data}
\title{Update raw data}
\usage{
update_raw_data(
  ds_tb,
  assignment_lup = NULL,
  correspondences_r3 = ready4show_correspondences(),
  datestamp_chr = character(0),
  drop_date_copy_1L_lgl = F,
  drop_time_1L_lgl = F,
  force_dates_1L_lgl = F,
  force_integers_1L_lgl = F,
  integers_chr = character(0),
  match_var_nm_1L_chr = character(0),
  non_integers_chr = character(0),
  recode_lgl_chr = character(0),
  recode_lgl_fn = NULL,
  recode_vals_fn_ls = NULL,
  response_id_var_nm_1L_chr = character(0)
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{assignment_lup}{Assignment (a lookup table), Default: NULL}

\item{correspondences_r3}{Correspondences (a ready4 submodule), Default: ready4show_correspondences()}

\item{datestamp_chr}{Datestamp (a character vector), Default: character(0)}

\item{drop_date_copy_1L_lgl}{Drop date copy (a logical vector of length one), Default: F}

\item{drop_time_1L_lgl}{Drop time (a logical vector of length one), Default: F}

\item{force_dates_1L_lgl}{Force dates (a logical vector of length one), Default: F}

\item{force_integers_1L_lgl}{Force integers (a logical vector of length one), Default: F}

\item{integers_chr}{Integers (a character vector), Default: character(0)}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: character(0)}

\item{non_integers_chr}{Non integers (a character vector), Default: character(0)}

\item{recode_lgl_chr}{Recode logical vector (a character vector), Default: character(0)}

\item{recode_lgl_fn}{Recode logical vector (a function), Default: NULL}

\item{recode_vals_fn_ls}{Recode values (a list of functions), Default: NULL}

\item{response_id_var_nm_1L_chr}{Response identity variable name (a character vector of length one), Default: character(0)}
}
\value{
Dataset (a tibble)
}
\description{
update_raw_data() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update raw data. The function returns Dataset (a tibble).
}
\keyword{internal}
