% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_update_src_loc_to_url.R
\name{update_src_loc_to_url}
\alias{update_src_loc_to_url}
\alias{update_src_loc_to_url.ready4_all_import_lup}
\alias{update_src_loc_to_url,ready4_all_import_lup-method}
\title{Update source local to url}
\usage{
update_src_loc_to_url(x, local_to_url_vec, urls_vec, ...)

update_src_loc_to_url.ready4_all_import_lup(x, local_to_url_vec, urls_vec)

\S4method{update_src_loc_to_url}{ready4_all_import_lup}(x, local_to_url_vec, urls_vec)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{local_to_url_vec}{PARAM_DESCRIPTION}

\item{urls_vec}{PARAM_DESCRIPTION}

\item{...}{Additional arguments}
}
\description{
update_src_loc_to_url() is an Update Source Local to Url generic that updates data from a local file reference to a URL

update_src_loc_to_url.ready4_all_import_lup() is an Update Source Local to Url method that updates data from a local file reference to a URL This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of sources of raw (un-processed) data to import.. The function is called for its side effects and does not return a value.
}
