% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_tb_src_loc_to_url_sgl_tb}
\alias{update_tb_src_loc_to_url_sgl_tb}
\title{Update tibble source local to url sgl}
\usage{
update_tb_src_loc_to_url_sgl_tb(x, y, local_to_url_vec, urls_vec)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{local_to_url_vec}{PARAM_DESCRIPTION}

\item{urls_vec}{PARAM_DESCRIPTION}
}
\value{
Updated (a tibble)
}
\description{
update_tb_src_loc_to_url_sgl_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update tibble source local to url sgl tibble. Function argument x specifies the object to be updated. Argument y provides the object to be updated. The function returns Updated (a tibble).
}
