% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_update.R
\name{update_tb_src_loc_to_url_sngl_tb}
\alias{update_tb_src_loc_to_url_sngl_tb}
\title{Update tibble source local to url single tibble}
\usage{
update_tb_src_loc_to_url_sngl_tb(x, y, local_to_url_vec_chr, urls_vec_chr)
}
\arguments{
\item{x}{An object}

\item{y}{An object}

\item{local_to_url_vec_chr}{Local to url vector (a character vector)}

\item{urls_vec_chr}{Urls vector (a character vector)}
}
\value{
Updated (a tibble)
}
\description{
update_tb_src_loc_to_url_sngl_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update tibble source local to url single tibble. The function returns Updated (a tibble).
}
\keyword{internal}
