% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{update_this}
\alias{update_this}
\title{Update this}
\usage{
update_this(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (an additional arguments)}
}
\description{
update_this() is an Update this generic that updates and object
}
