% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dictionary.R
\name{validate_ready4_dictionary}
\alias{validate_ready4_dictionary}
\title{Validate ready4 s3 class defining a data dictionary tibble.}
\usage{
validate_ready4_dictionary(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 s3 class defining a data dictionary tibble.}
}
\value{
A prototpe for ready4 s3 class defining a data dictionary tibble.
}
\description{
Validate an instance of the ready4 s3 class defining a data dictionary tibble.
}
\details{
ready4 s3 class defining a data dictionary tibble.
}
