% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_struc_mape.R
\name{validate_ready4_par_struc_mape}
\alias{validate_ready4_par_struc_mape}
\title{Validate ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
validate_ready4_par_struc_mape(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
}
\value{
A prototpe for ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Validate an instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
