% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_dictionary.R
\name{validate_ready4use_dictionary}
\alias{validate_ready4use_dictionary}
\title{validate ready4use package dictionary ready4 s3 class defining a data dictionary tibble.}
\usage{
validate_ready4use_dictionary(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 s3 class defining a data dictionary tibble.}
}
\value{
A prototpe for ready4 s3 class defining a data dictionary tibble.
}
\description{
Validate an instance of the ready4 s3 class defining a data dictionary tibble.
}
\details{
A data dictionary tibble.
}
\keyword{internal}
