% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_dist.R
\name{validate_ready4use_dist}
\alias{validate_ready4use_dist}
\title{Validate ready4use dist ready4 S3 class for list object that summarises the parameters of each distribution}
\usage{
validate_ready4use_dist(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 S3 class for list object that summarises the parameters of each distribution}
}
\value{
A prototpe for ready4 S3 class for list object that summarises the parameters of each distribution
}
\description{
Validate an instance of the ready4 S3 class for list object that summarises the parameters of each distribution
}
\details{
ready4 S3 class for list object that summarises the parameters of each distribution
}
