% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_manifest.R
\name{validate_ready4use_manifest}
\alias{validate_ready4use_manifest}
\title{Validate ready4use package manifest ready4 s3 class defining a manifest of data required to create an R package.}
\usage{
validate_ready4use_manifest(x)
}
\arguments{
\item{x}{An unvalidated instance of the ready4 s3 class defining a manifest of data required to create an R package.}
}
\value{
A prototpe for ready4 s3 class defining a manifest of data required to create an R package.
}
\description{
Validate an instance of the ready4 s3 class defining a manifest of data required to create an R package.
}
\details{
ready4 s3 class Manifest for packages containing datasets.
}
