% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_dv_ds}
\alias{write_dv_ds}
\title{write dataverse dataset}
\usage{
write_dv_ds(
  ds_meta_ls,
  dev_pkg_nm_1L_chr = deprecated(),
  dss_tb,
  dv_nm_1L_chr,
  parent_dv_dir_1L_chr,
  paths_to_dirs_chr,
  inc_fl_types_chr = NA_character_,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{ds_meta_ls}{Dataset meta (a list)}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: deprecated()}

\item{dss_tb}{Datasets (a tibble)}

\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one)}

\item{parent_dv_dir_1L_chr}{Parent dataverse directory (a character vector of length one)}

\item{paths_to_dirs_chr}{Paths to directories (a character vector)}

\item{inc_fl_types_chr}{Include file types (a character vector), Default: 'NA'}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Dataset (a list)
}
\description{
write_dv_ds() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write dataverse dataset. The function returns Dataset (a list).
}
