% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_paired_ds_fls_to_dv}
\alias{write_paired_ds_fls_to_dv}
\title{Write paired dataset files to dataverse}
\usage{
write_paired_ds_fls_to_dv(
  ds_tb,
  fl_nm_1L_chr,
  desc_1L_chr,
  ds_url_1L_chr = "https://doi.org/10.7910/DVN/2Y9VF9",
  pkg_dv_dir_1L_chr = "data-raw/dataverse",
  data_dir_rt_1L_chr = ".",
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{desc_1L_chr}{Description (a character vector of length one)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one), Default: 'https://doi.org/10.7910/DVN/2Y9VF9'}

\item{pkg_dv_dir_1L_chr}{Package dataverse directory (a character vector of length one), Default: 'data-raw/dataverse'}

\item{data_dir_rt_1L_chr}{Data directory root (a character vector of length one), Default: '.'}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\description{
write_paired_ds_fls_to_dv() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write paired dataset files to dataverse. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
