% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_pkg_dss_to_dv_ds_csvs}
\alias{write_pkg_dss_to_dv_ds_csvs}
\title{Write package datasets to dataverse dataset comma separated variables files}
\usage{
write_pkg_dss_to_dv_ds_csvs(
  pkg_dss_tb,
  dv_nm_1L_chr,
  ds_url_1L_chr,
  wait_time_in_secs_int = 5L,
  dev_pkg_nm_1L_chr = ready4fun::get_dev_pkg_nm(),
  parent_dv_dir_1L_chr = "../../../../Data/Dataverse",
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER")
)
}
\arguments{
\item{pkg_dss_tb}{Package datasets (a tibble)}

\item{dv_nm_1L_chr}{Dataverse name (a character vector of length one)}

\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{wait_time_in_secs_int}{Wait time in secs (an integer vector), Default: 5}

\item{dev_pkg_nm_1L_chr}{Development package name (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{parent_dv_dir_1L_chr}{Parent dataverse directory (a character vector of length one), Default: '../../../../Data/Dataverse'}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}
}
\value{
Dataset (a list)
}
\description{
write_pkg_dss_to_dv_ds_csvs() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write package datasets to dataverse dataset comma separated variables files. The function returns Dataset (a list).
}
\keyword{internal}
