% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_add_urls_to_dss}
\alias{write_to_add_urls_to_dss}
\title{Write to add urls to datasets}
\usage{
write_to_add_urls_to_dss(
  ds_url_1L_chr,
  pkg_dss_tb,
  pkg_nm_1L_chr = ready4fun::get_dev_pkg_nm(),
  object_type_lup = NULL
)
}
\arguments{
\item{ds_url_1L_chr}{Dataset url (a character vector of length one)}

\item{pkg_dss_tb}{Package datasets (a tibble)}

\item{pkg_nm_1L_chr}{Package name (a character vector of length one), Default: ready4fun::get_dev_pkg_nm()}

\item{object_type_lup}{Object type (a lookup table), Default: NULL}
}
\value{
Package datasets (a tibble)
}
\description{
write_to_add_urls_to_dss() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to add urls to datasets. The function returns Package datasets (a tibble).
}
\keyword{internal}
