% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_to_copy_fls_to_dv_dir}
\alias{write_to_copy_fls_to_dv_dir}
\title{Write to copy files to dataverse directory}
\usage{
write_to_copy_fls_to_dv_dir(files_tb, local_dv_dir_1L_chr)
}
\arguments{
\item{files_tb}{Files (a tibble)}

\item{local_dv_dir_1L_chr}{Local dataverse directory (a character vector of length one)}
}
\description{
write_to_copy_fls_to_dv_dir() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write to copy files to dataverse directory. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour
}
