% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_write_type_1L_chr.R
\name{write_type_1L_chr}
\alias{write_type_1L_chr}
\alias{write_type_1L_chr-Ready4useRaw}
\alias{write_type_1L_chr,Ready4useRaw-method}
\alias{write_type_1L_chr-Ready4useProcessed}
\alias{write_type_1L_chr,Ready4useProcessed-method}
\title{write_type_1L_chr}
\usage{
write_type_1L_chr(x)

\S4method{write_type_1L_chr}{Ready4useRaw}(x)

\S4method{write_type_1L_chr}{Ready4useProcessed}(x)
}
\arguments{
\item{x}{An object of class Ready4useProcessed}
}
\description{
S4 Generic function to get the value of the slot write_type_1L_chr

Get the value of the slot write_type_1L_chr for S4 objects of class Ready4useRaw

Get the value of the slot write_type_1L_chr for S4 objects of class Ready4useProcessed
}
