% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_write_type_1L_chr.R
\name{write_type_1L_chr<-}
\alias{write_type_1L_chr<-}
\alias{write_type_1L_chr<--Ready4useRaw}
\alias{write_type_1L_chr<-,Ready4useRaw-method}
\alias{write_type_1L_chr<--Ready4useProcessed}
\alias{write_type_1L_chr<-,Ready4useProcessed-method}
\title{write_type_1L_chr<-}
\usage{
write_type_1L_chr(x) <- value

\S4method{write_type_1L_chr}{Ready4useRaw}(x) <- value

\S4method{write_type_1L_chr}{Ready4useProcessed}(x) <- value
}
\arguments{
\item{x}{An object of class Ready4useProcessed}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot write_type_1L_chr

Set the value of the slot write_type_1L_chr for S4 objects of class Ready4useRaw

Set the value of the slot write_type_1L_chr for S4 objects of class Ready4useProcessed
}
