% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_labels_from_dictionary}
\alias{add_labels_from_dictionary}
\title{Add labels from dictionary}
\usage{
add_labels_from_dictionary(ds_tb, dictionary_tb, remove_old_lbls_1L_lgl = F)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{dictionary_tb}{Dictionary (a tibble)}

\item{remove_old_lbls_1L_lgl}{Remove old lbls (a logical vector of length one), Default: F}
}
\value{
Labelled dataset (a tibble)
}
\description{
add_labels_from_dictionary() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add labels from dictionary. Function argument ds_tb specifies the object to be updated. The function returns Labelled dataset (a tibble).
}
