% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_bind_lups.R
\name{bind_lups}
\alias{bind_lups}
\alias{bind_lups.ready4_dictionary}
\alias{bind_lups,ready4_dictionary-method}
\title{Bind lups}
\usage{
bind_lups(x, ...)

bind_lups.ready4_dictionary(x, new_ready4_dict_r3)

\S4method{bind_lups}{ready4_dictionary}(x, new_ready4_dict_r3)
}
\arguments{
\item{x}{An instance of ready4 s3 class defining a data dictionary tibble.}

\item{...}{Additional arguments}

\item{new_ready4_dict_r3}{New ready4 dict (a ready4 S3)}
}
\value{
NA ()
}
\description{
bind_lups() is a Bind Lups generic that rowbinds lookup tables of the same class, removing duplicates based on priority.

bind_lups.ready4_dictionary() is a Bind Lups method that rowbinds lookup tables of the same class, removing duplicates based on priority. This method is implemented for the ready4 s3 class defining a data dictionary tibble.. The function is called for its side effects and does not return a value.
}
