% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_crs_nbr_dbl.R
\name{crs_nbr_dbl-ready4_script_data}
\alias{crs_nbr_dbl-ready4_script_data}
\alias{crs_nbr_dbl,ready4_script_data-method}
\title{crs_nbr_dbl}
\usage{
\S4method{crs_nbr_dbl}{ready4_script_data}(x)
}
\arguments{
\item{x}{An object of class ready4_script_data}
}
\description{
Get the value of the slot crs_nbr_dbl for S4 objects of class ready4_script_data
}
