% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_crs_nbr_dbl.R
\name{crs_nbr_dbl}
\alias{crs_nbr_dbl}
\alias{crs_nbr_dbl-Ready4useArguments}
\alias{crs_nbr_dbl,Ready4useArguments-method}
\title{crs_nbr_dbl}
\usage{
crs_nbr_dbl(x)

\S4method{crs_nbr_dbl}{Ready4useArguments}(x)
}
\arguments{
\item{x}{An object of class Ready4useArguments}
}
\description{
S4 Generic function to get the value of the slot crs_nbr_dbl

Get the value of the slot crs_nbr_dbl for S4 objects of class Ready4useArguments
}
