% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrc_crs_nbr_dbl.R
\name{crs_nbr_dbl<-}
\alias{crs_nbr_dbl<-}
\alias{crs_nbr_dbl<--ready4_script_data}
\alias{crs_nbr_dbl<-,ready4_script_data-method}
\title{crs_nbr_dbl<-}
\usage{
crs_nbr_dbl(x) <- value

\S4method{crs_nbr_dbl}{ready4_script_data}(x) <- value
}
\arguments{
\item{x}{An object of class ready4_script_data}

\item{value}{Value to be assigned to x}
}
\description{
S4 Generic function to set the value of the slot crs_nbr_dbl

Set the value of the slot crs_nbr_dbl for S4 objects of class ready4_script_data
}
