% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exhibit.R
\name{exhibit.ready4use_dictionary}
\alias{exhibit.ready4use_dictionary}
\alias{exhibit,ready4use_dictionary-method}
\alias{exhibit-Ready4useDyad}
\alias{exhibit,Ready4useDyad-method}
\title{Exhibit features of model module data by printing them to the R console}
\usage{
exhibit.ready4use_dictionary(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{ready4use_dictionary}(
  x,
  caption_1L_chr = NULL,
  mkdn_tbl_ref_1L_chr = NULL,
  output_type_1L_chr = "HTML",
  use_lbls_as_col_nms_1L_lgl = T
)

\S4method{exhibit}{Ready4useDyad}(
  x,
  caption_1L_chr = NA_character_,
  display_1L_chr = "all",
  label_1L_lgl = FALSE,
  mkdn_tbl_ref_1L_chr = "",
  output_type_1L_chr = "HTML",
  type_1L_chr = "ds",
  use_lbls_as_col_nms_1L_lgl = TRUE,
  use_rdocx_1L_lgl = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class Ready4useDyad}

\item{caption_1L_chr}{Caption (a character vector of length one), Default: 'NA'}

\item{mkdn_tbl_ref_1L_chr}{Markdown table reference (a character vector of length one), Default: ''}

\item{output_type_1L_chr}{Output type (a character vector of length one), Default: 'HTML'}

\item{use_lbls_as_col_nms_1L_lgl}{Use labels as column names (a logical vector of length one), Default: TRUE}

\item{display_1L_chr}{Display (a character vector of length one), Default: 'all'}

\item{label_1L_lgl}{Label (a logical vector of length one), Default: FALSE}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'ds'}

\item{use_rdocx_1L_lgl}{Use rdocx (a logical vector of length one), Default: FALSE}

\item{...}{Additional arguments}
}
\value{
No return value, called for side effects.

No return value, called for side effects.
}
\description{
exhibit.ready4use_dictionary() is an exhibit method that exhibits features of a class instance by printing to console. This method is implemented for the ready4 s3 class defining a data dictionary tibble. The function is called for its side effects and does not return a value.

exhibit method applied to Ready4useDyad
}
