% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_file_from_dv}
\alias{get_file_from_dv}
\title{Get file from dataverse}
\usage{
get_file_from_dv(
  database_ui_chr,
  filename_chr,
  save_format_chr,
  repo_file_format,
  dataverse_chr = Sys.getenv("DATAVERSE_SERVER"),
  save_type_chr = "original",
  save_dir_path_chr = "",
  read_fn,
  unlink_lgl = T
)
}
\arguments{
\item{database_ui_chr}{Database ui (a character vector of length 1)}

\item{filename_chr}{Filename (a character vector of length 1)}

\item{save_format_chr}{Save format (a character vector of length 1)}

\item{repo_file_format}{PARAM_DESCRIPTION}

\item{dataverse_chr}{Dataverse (a character vector of length 1), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{save_type_chr}{Save type (a character vector of length 1), Default: 'original'}

\item{save_dir_path_chr}{Save directory path (a character vector of length 1), Default: ''}

\item{read_fn}{Read (a function)}

\item{unlink_lgl}{Unlink (a logical vector of length 1), Default: T}
}
\description{
get_file_from_dv() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get a file from dv. Function argument database_ui_chr specifies the where to look for the required object.NA
}
\keyword{internal}
