% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_local_path_to_dv_data}
\alias{get_local_path_to_dv_data}
\title{Get local path to dataverse data}
\usage{
get_local_path_to_dv_data(save_dir_path_chr, filename_chr, save_format_chr)
}
\arguments{
\item{save_dir_path_chr}{Save directory path (a character vector)}

\item{filename_chr}{Filename (a character vector)}

\item{save_format_chr}{Save format (a character vector)}
}
\value{
Path (a character vector)
}
\description{
get_local_path_to_dv_data() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get local path to dataverse data. Function argument save_dir_path_chr specifies the where to look for the required object. The function returns Path (a character vector).
}
\keyword{internal}
