% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_local_path_to_dv_data}
\alias{get_local_path_to_dv_data}
\title{Get local path to dataverse data}
\usage{
get_local_path_to_dv_data(save_dir_path_1L_chr, fl_nm_1L_chr, save_fmt_1L_chr)
}
\arguments{
\item{save_dir_path_1L_chr}{Save directory path (a character vector of length one)}

\item{fl_nm_1L_chr}{File name (a character vector of length one)}

\item{save_fmt_1L_chr}{Save format (a character vector of length one)}
}
\value{
Path (a character vector)
}
\description{
get_local_path_to_dv_data() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get local path to dataverse data. Function argument save_dir_path_1L_chr specifies the where to look for the required object. The function returns Path (a character vector).
}
\keyword{internal}
