% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_r3_from_dv_csv}
\alias{get_r3_from_dv_csv}
\title{get ready4 S3 from dataverse comma separated variables file}
\usage{
get_r3_from_dv_csv(
  file_name_chr,
  data_repo_db_ui_chr,
  data_repo_ui_chr = NA_character_,
  r3_fn = ready4use_imports
)
}
\arguments{
\item{file_name_chr}{File name (a character vector)}

\item{data_repo_db_ui_chr}{Data repository database user interface (a character vector)}

\item{data_repo_ui_chr}{Data repository user interface (a character vector), Default: 'NA'}

\item{r3_fn}{Ready4 S3 (a function), Default: ready4use_imports}
}
\value{
Tibble ready4 S3 (a ready4 S3 extension of tibble)
}
\description{
get_r3_from_dv_csv() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get ready4 s3 from dataverse comma separated variables file. Function argument file_name_chr specifies the where to look for the required object. The function returns Tibble ready4 S3 (a ready4 S3 extension of tibble).
}
