% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R, R/mthd_get_read_fn.R
\name{get_read_fn}
\alias{get_read_fn}
\alias{get_read_fn.ready4_dv_import_lup}
\alias{get_read_fn,ready4_dv_import_lup-method}
\title{Get read function}
\usage{
get_read_fn(x, ...)

get_read_fn.ready4_dv_import_lup(x)

\S4method{get_read_fn}{ready4_dv_import_lup}(x)
}
\arguments{
\item{x}{An instance of Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.}

\item{...}{Additional arguments}
}
\description{
get_read_fn() is a Get Read Function generic that retrieves a read function.

get_read_fn.ready4_dv_import_lup() is a Get Read Function method that retrieves a read function. This method is implemented for the Readyforwhatsnext S3 class for tibble object lookup table of files to be imported from a dataverse.. The function is called for its side effects and does not return a value.
}
