% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ingest.R
\name{ingest-Ready4useRepos}
\alias{ingest-Ready4useRepos}
\alias{ingest,Ready4useRepos-method}
\title{Ingest data saved in external files into R objects stored in working memory}
\usage{
\S4method{ingest}{Ready4useRepos}(
  x,
  fls_to_ingest_chr = NA_character_,
  key_1L_chr = NULL,
  type_1L_chr = "R"
)
}
\arguments{
\item{x}{An object of class Ready4useRepos}

\item{fls_to_ingest_chr}{Files to ingest (a character vector), Default: 'NA'}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'R'}
}
\value{
X (Ingested data, descriptive metadata and provenance details.)
}
\description{
ingest method applied to Ready4useRepos
}
