% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_ingest.R
\name{ingest-Ready4useRepos}
\alias{ingest-Ready4useRepos}
\alias{ingest,Ready4useRepos-method}
\title{Ingest data}
\usage{
\S4method{ingest}{Ready4useRepos}(
  x,
  fls_to_ingest_chr = NA_character_,
  gh_token_1L_chr = "",
  key_1L_chr = NULL,
  metadata_1L_lgl = T,
  type_1L_chr = "R"
)
}
\arguments{
\item{x}{An object of class Ready4useRepos}

\item{fls_to_ingest_chr}{Files to ingest (a character vector), Default: 'NA'}

\item{gh_token_1L_chr}{Github token (a character vector of length one), Default: ''}

\item{key_1L_chr}{Key (a character vector of length one), Default: NULL}

\item{metadata_1L_lgl}{Metadata (a logical vector of length one), Default: T}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'R'}
}
\value{
Ingest (an output object of multiple potential types)
}
\description{
ingest method applied to Ready4useRepos
}
