% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dictionary.R
\name{make_pt_ready4_dictionary}
\alias{make_pt_ready4_dictionary}
\title{Make prototype ready4 s3 class defining a data dictionary tibble.}
\usage{
make_pt_ready4_dictionary(
  var_nm_chr = character(0),
  var_ctg_chr = character(0),
  var_desc_chr = character(0),
  var_type_chr = character(0)
)
}
\arguments{
\item{var_nm_chr}{Variable name (a character vector), Default: character(0)}

\item{var_ctg_chr}{Variable category categories (a character vector), Default: character(0)}

\item{var_desc_chr}{Variable description (a character vector), Default: character(0)}

\item{var_type_chr}{Variable type (a character vector), Default: character(0)}
}
\value{
A prototype for ready4 s3 class defining a data dictionary tibble.
}
\description{
Create a new prototype for the ready4 s3 class defining a data dictionary tibble.
}
\details{
ready4 s3 class defining a data dictionary tibble.
}
