% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dist.R
\name{make_pt_ready4_dist}
\alias{make_pt_ready4_dist}
\title{Make prototype ready4 S3 class for list object that summarises the parameters of each distribution}
\usage{
make_pt_ready4_dist(
  distribution_chr = character(0),
  dist_param_1_dbl = numeric(0),
  dist_param_2_dbl = numeric(0),
  dist_param_3_dbl = numeric(0),
  dist_param_4_dbl = numeric(0),
  transformation_chr = character(0)
)
}
\arguments{
\item{distribution_chr}{Distribution (a character vector), Default: character(0)}

\item{dist_param_1_dbl}{Dist param 1 (a double vector), Default: numeric(0)}

\item{dist_param_2_dbl}{Dist param 2 (a double vector), Default: numeric(0)}

\item{dist_param_3_dbl}{Dist param 3 (a double vector), Default: numeric(0)}

\item{dist_param_4_dbl}{Dist param 4 (a double vector), Default: numeric(0)}

\item{transformation_chr}{Transformation (a character vector), Default: character(0)}
}
\value{
A prototype for ready4 S3 class for list object that summarises the parameters of each distribution
}
\description{
Create a new prototype for the ready4 S3 class for list object that summarises the parameters of each distribution
}
\details{
ready4 S3 class for list object that summarises the parameters of each distribution
}
