% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_dist.R
\name{make_pt_ready4_dist}
\alias{make_pt_ready4_dist}
\title{Make prototype ready4 S3 class for list object that summarises the parameters of each distribution}
\usage{
make_pt_ready4_dist(
  distribution = character(0),
  dist_param_1 = numeric(0),
  dist_param_2 = numeric(0),
  dist_param_3 = numeric(0),
  dist_param_4 = numeric(0),
  transformation = character(0)
)
}
\arguments{
\item{distribution}{PARAM_DESCRIPTION, Default: character(0)}

\item{dist_param_1}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{dist_param_2}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{dist_param_3}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{dist_param_4}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{transformation}{PARAM_DESCRIPTION, Default: character(0)}
}
\value{
A prototype for
}
\description{
Create a new prototype for the
}
\details{
ready4 S3 class for list object that summarises the parameters of each distribution
}
