% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4_par_struc_mape.R
\name{make_pt_ready4_par_struc_mape}
\alias{make_pt_ready4_par_struc_mape}
\title{Make prototype ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.}
\usage{
make_pt_ready4_par_struc_mape(
  param_name = character(0),
  sex_age_band = character(0),
  mape_05_yr_mde = numeric(0),
  mape_10_yr_mde = numeric(0),
  mape_15_yr_mde = numeric(0),
  mape_05_yr_min = numeric(0),
  mape_10_yr_min = numeric(0),
  mape_15_yr_min = numeric(0),
  mape_05_yr_max = numeric(0),
  mape_10_yr_max = numeric(0),
  mape_15_yr_max = numeric(0),
  mape_05_yr_shp = numeric(0),
  mape_10_yr_shp = numeric(0),
  mape_15_yr_shp = numeric(0)
)
}
\arguments{
\item{param_name}{PARAM_DESCRIPTION, Default: character(0)}

\item{sex_age_band}{PARAM_DESCRIPTION, Default: character(0)}

\item{mape_05_yr_mde}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_10_yr_mde}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_15_yr_mde}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_05_yr_min}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_10_yr_min}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_15_yr_min}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_05_yr_max}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_10_yr_max}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_15_yr_max}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_05_yr_shp}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_10_yr_shp}{PARAM_DESCRIPTION, Default: numeric(0)}

\item{mape_15_yr_shp}{PARAM_DESCRIPTION, Default: numeric(0)}
}
\value{
A prototype for ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\description{
Create a new prototype for the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
\details{
ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.
}
