% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_r3_from_csv_tb}
\alias{make_r3_from_csv_tb}
\title{make ready4 S3 from comma separated variables file tibble}
\usage{
make_r3_from_csv_tb(csv_tb, r3_fn)
}
\arguments{
\item{csv_tb}{Comma separated variables file (a tibble)}

\item{r3_fn}{Ready4 S3 (a function)}
}
\value{
Tibble ready4 S3 (a ready4 S3 extension of tibble)
}
\description{
make_r3_from_csv_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make ready4 s3 from comma separated variables file tibble. The function returns Tibble ready4 S3 (a ready4 S3 extension of tibble).
}
\keyword{internal}
