% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture.ready4use_dataverses}
\alias{manufacture.ready4use_dataverses}
\alias{manufacture,ready4use_dataverses-method}
\alias{manufacture.ready4use_imports}
\alias{manufacture,ready4use_imports-method}
\title{manufacture - a method that manufactures a novel r object using data contained in an instance of a class}
\usage{
manufacture.ready4use_dataverses(x, type_1L_chr = "read_fn")

\S4method{manufacture}{ready4use_dataverses}(x, type_1L_chr = "read_fn")

manufacture.ready4use_imports(
  x,
  forced_choice_chr = NA_character_,
  script_args_ls = NULL
)

\S4method{manufacture}{ready4use_imports}(x, forced_choice_chr = NA_character_, script_args_ls = NULL)
}
\arguments{
\item{x}{An instance of ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{type_1L_chr}{Type (a character vector of length one), Default: 'read_fn'}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{script_args_ls}{Script arguments (a list), Default: NULL}
}
\value{
Read (a list of functions)
}
\description{
manufacture.ready4use_dataverses() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of files to be imported from a dataverse. The function returns Read (a list of functions).

manufacture.ready4use_imports() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.
}
