% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_manufacture.R
\name{manufacture.ready4use_imports}
\alias{manufacture.ready4use_imports}
\alias{manufacture,ready4use_imports-method}
\title{Manufacture - a method that creates a novel r object.}
\usage{
manufacture.ready4use_imports(
  x,
  forced_choice_chr = NA_character_,
  script_args_ls = NULL
)

\S4method{manufacture}{ready4use_imports}(x, forced_choice_chr = NA_character_, script_args_ls = NULL)
}
\arguments{
\item{x}{An instance of ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import.}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{script_args_ls}{Script arguments (a list), Default: NULL}
}
\description{
manufacture.ready4use_imports() is a Manufacture method that creates a novel R object. This method is implemented for the ready4 S3 class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.
}
