% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_procure.R
\name{procure.ready4use_dataverses}
\alias{procure.ready4use_dataverses}
\alias{procure,ready4use_dataverses-method}
\alias{procure.ready4use_imports}
\alias{procure,ready4use_imports-method}
\alias{procure-Ready4useIngest}
\alias{procure,Ready4useIngest-method}
\title{Procure items from a dataset}
\usage{
procure.ready4use_dataverses(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

\S4method{procure}{ready4use_dataverses}(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

procure.ready4use_imports(
  x,
  inc_script_1L_lgl = T,
  forced_choice_chr = NA_character_
)

\S4method{procure}{ready4use_imports}(x, inc_script_1L_lgl = T, forced_choice_chr = NA_character_)

\S4method{procure}{Ready4useIngest}(x, fl_nm_1L_chr = NA_character_)
}
\arguments{
\item{x}{An object of class Ready4useIngest}

\item{save_dir_path_1L_chr}{Save directory path (a character vector of length one), Default: ''}

\item{unlink_1L_lgl}{Unlink (a logical vector of length one), Default: T}

\item{server_1L_chr}{Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{key_1L_chr}{Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")}

\item{inc_script_1L_lgl}{Include script (a logical vector of length one), Default: T}

\item{forced_choice_chr}{Forced choice (a character vector), Default: 'NA'}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'NA'}
}
\value{
Data (an output object of multiple potential types)

No return value, called for side effects.

Object (an output object of multiple potential types)
}
\description{
procure.ready4use_dataverses() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of files to be imported from a dataverse. The function returns Data (an output object of multiple potential types).

procure.ready4use_imports() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.

procure method applied to Ready4useIngest
}
