% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_read.R
\name{read_import_from_csv}
\alias{read_import_from_csv}
\title{Read import from csv}
\usage{
read_import_from_csv(file_ref_chr, is_url = F)
}
\arguments{
\item{file_ref_chr}{File ref (a character vector of length 1)}

\item{is_url}{PARAM_DESCRIPTION, Default: F}
}
\value{
Import (a readyforwhatsnext S3)
}
\description{
read_import_from_csv() is a Read function that reads an R script into memory. Specifically, this function implements an algorithm to read an import from csv. Function argument file_ref_chr specifies the path to object.The function returns an import (a readyforwhatsnext s3).
}
\keyword{internal}
