% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_read.R
\name{read_import_from_csv}
\alias{read_import_from_csv}
\title{Read import from comma separated variables file}
\usage{
read_import_from_csv(file_ref_chr, is_url = F)
}
\arguments{
\item{file_ref_chr}{File reference (a character vector)}

\item{is_url}{PARAM_DESCRIPTION, Default: F}
}
\value{
Import (a ready4 S3)
}
\description{
read_import_from_csv() is a Read function that reads an R script into memory. Specifically, this function implements an algorithm to read import from comma separated variables file. Function argument file_ref_chr specifies the path to object. The function returns Import (a ready4 S3).
}
\keyword{internal}
