% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_local.R
\docType{class}
\name{ready4_local}
\alias{ready4_local}
\title{ready4_local}
\usage{
ready4_local(
  merge_with_chr_vec = NA_character_,
  raw_data_dir_chr = NA_character_,
  pckg_chr = NA_character_,
  overwrite_lgl = NA,
  save_lgl = NA
)
}
\arguments{
\item{merge_with_chr_vec}{character, Default: 'NA'}

\item{raw_data_dir_chr}{character, Default: 'NA'}

\item{pckg_chr}{character, Default: 'NA'}

\item{overwrite_lgl}{logical, Default: NA}

\item{save_lgl}{logical, Default: NA}
}
\value{
An S4 object of the ready4_local class
}
\description{
An S4 class to represent Object defining data to be saved in local directory.

Create a new S4 object of the class:ready4_local
}
\section{Slots}{

\describe{
\item{\code{merge_with_chr_vec}}{character}

\item{\code{raw_data_dir_chr}}{character}

\item{\code{pckg_chr}}{character}

\item{\code{overwrite_lgl}}{logical}

\item{\code{save_lgl}}{logical}
}}

