% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_local.R
\docType{class}
\name{ready4_local}
\alias{ready4_local}
\title{ready4_local}
\description{
An S4 class to represent ready4 S4 class defining data to be saved in local directory.
}
\section{Slots}{

\describe{
\item{\code{merge_with_chr_vec}}{character}
}}

