% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_local_proc.R
\docType{class}
\name{ready4_local_proc}
\alias{ready4_local_proc}
\title{ready4_local_proc}
\description{
An S4 class to represent ready4 S4 class defining data to be saved in local directory in a processed (R) format.
}
\section{Slots}{

\describe{
\item{\code{save_type}}{character}

\item{\code{merge_with_chr_vec}}{character}
}}

