% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_local_raw.R
\docType{class}
\name{ready4_local_raw}
\alias{ready4_local_raw}
\title{ready4_local_raw}
\description{
An S4 class to represent Readyforwhatsnext S4 class defining data to be saved in local directory in a raw (unprocessed) format.
}
\section{Slots}{

\describe{
\item{\code{save_type}}{character}

\item{\code{merge_with_chr_vec}}{character}
}}

