% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_ready4_script_data.R
\docType{class}
\name{ready4_script_data}
\alias{ready4_script_data}
\title{ready4_script_data}
\usage{
ready4_script_data(
  save_type = NA_character_,
  merge_with_chr_vec = NA_character_,
  crs_nbr_vec = NA_real_
)
}
\arguments{
\item{save_type}{character, Default: 'NA'}

\item{merge_with_chr_vec}{character, Default: 'NA'}

\item{crs_nbr_vec}{numeric, Default: NA}
}
\value{
An S4 object of the
}
\description{
An S4 class to represent Readyforwhatsnext S4 class containint data to be passed to a function that constructs a spatial object from a lookup table.

Create a new S4 object of the class:ready4_script_data
}
\section{Slots}{

\describe{
\item{\code{crs_nbr_vec}}{numeric}

\item{\code{save_type}}{character}

\item{\code{merge_with_chr_vec}}{character}
}}

