% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4use.R
\docType{package}
\name{ready4use-package}
\alias{ready4use}
\alias{ready4use-package}
\title{ready4use: Retrieve and Manage Data in Open and Modular Mental Health Simulations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

ready4use provides a set of tools for general data management tasks when developping open source, modular mental health simulation models. This development version of the ready4use package has been made available as part of the process of testing and documenting the package. The tools contained in this development release automate a number of tasks which MODIFY THE DIRECTORY STRUCTURE OF YOUR LOCAL MACHINE. Therefore you should only trial this software if you feel confident that you understand what it does and have created a sandpit area in which you can safely undertake testing. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4use/}
  \item \url{https://github.com/ready4-dev/ready4use}
  \item \url{https://ready4-dev.github.io/ready4/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
