% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4use.R, R/ready4use-package.R
\docType{package}
\name{ready4use-package}
\alias{ready4use}
\alias{ready4use-package}
\title{ready4use: Author, Label and Share Ready4 Framework Datasets}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

ready4use provides a set of tools for authoring and working with datasets when developping open source, modular mental health systems models. This development version of the ready4use package has been made available as part of the process of testing and documenting the package. You should only trial this software if you feel confident that you understand what it does and have created a sandpit area in which you can safely undertake testing. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).

\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

ready4use provides a set of tools for authoring and working with datasets when developping open source, modular mental health systems models. This development version of the ready4use package has been made available as part of the process of testing and documenting the package. You should only trial this software if you feel confident that you understand what it does and have created a sandpit area in which you can safely undertake testing. If you have any questions, please contact the authors (matthew.hamilton@orygen.org.au).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4use/}
  \item \url{https://github.com/ready4-dev/ready4use}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/ready4use/}
  \item \url{https://github.com/ready4-dev/ready4use}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}


\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
\keyword{internal}
