% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ready4use.R
\docType{package}
\name{ready4use-package}
\alias{ready4use}
\alias{ready4use-package}
\title{ready4use: Readyforwhatsnext Methods for retrieving and
managing data.}
\description{
ready4use provides a set of classes and
methods for general data management tasks throughout the
readyforwhatsnext suite.
}
\details{
To learn more about ready4use, start with the vignettes:
\code{browseVignettes(package = "ready4use")}
}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton@orygen.org.au} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID})

Authors:
\itemize{
  \item Glen Wiesner \email{Glen.Wiesner@vu.edu.au} (\href{https://orcid.org/0000-0002-0071-130X}{ORCID})
}

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
  \item VicHealth [funder]
  \item Victoria University [funder]
}

}
