% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C3_ready4use_distributions.R
\name{ready4use_distributions}
\alias{ready4use_distributions}
\title{ready4 S3 class for list object that summarises the parameters of each distribution}
\usage{
ready4use_distributions(x = make_pt_ready4use_distributions())
}
\arguments{
\item{x}{A prototype for the ready4 S3 class for list object that summarises the parameters of each distribution, Default: make_pt_ready4use_distributions()}
}
\value{
A validated instance of the ready4 S3 class for list object that summarises the parameters of each distribution
}
\description{
Create a new valid instance of the ready4 S3 class for list object that summarises the parameters of each distribution
}
\details{
ready4 S3 class for list object that summarises the parameters of each distribution
}
