% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useRaw.R
\docType{class}
\name{Ready4useRaw-class}
\alias{Ready4useRaw-class}
\alias{Ready4useRaw}
\title{Ready4useRaw}
\description{
ready4 S4 class defining data to be saved in local directory in a raw (unprocessed) format.
}
\section{Slots}{

\describe{
\item{\code{write_type_1L_chr}}{Write type (a character vector of length one)}

\item{\code{merge_itms_chr}}{Merge items (a character vector)}

\item{\code{raw_fls_dir_1L_chr}}{Raw files directory (a character vector of length one)}

\item{\code{pkg_1L_chr}}{Package (a character vector of length one)}

\item{\code{overwrite_1L_lgl}}{Overwrite (a logical vector of length one)}

\item{\code{write_1L_lgl}}{Write (a logical vector of length one)}
}}

