% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4_Ready4useRecord.R
\docType{class}
\name{Ready4useRecord-class}
\alias{Ready4useRecord-class}
\alias{Ready4useRecord}
\title{Ready4useRecord}
\description{
Ingested data, descriptive metadata and provenance details.
}
\section{Slots}{

\describe{
\item{\code{a_Ready4usePointer}}{NO MATCH}

\item{\code{b_Ready4useIngest}}{NO MATCH}

\item{\code{dissemination_1L_chr}}{Dissemination (a character vector of length one)}
}}

