% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_renew.R
\name{renew.ready4use_dictionary}
\alias{renew.ready4use_dictionary}
\alias{renew,ready4use_dictionary-method}
\alias{renew.ready4use_imports}
\alias{renew,ready4use_imports-method}
\alias{renew-Ready4useDyad}
\alias{renew,Ready4useDyad-method}
\title{Renew (update) values}
\usage{
renew.ready4use_dictionary(
  x,
  var_nm_chr = NA_character_,
  var_ctg_chr = NA_character_,
  var_desc_chr = NA_character_,
  var_type_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

\S4method{renew}{ready4use_dictionary}(
  x,
  var_nm_chr = NA_character_,
  var_ctg_chr = NA_character_,
  var_desc_chr = NA_character_,
  var_type_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

renew.ready4use_imports(
  x,
  local_file_src_chr = NA_character_,
  path_to_make_script_chr = NA_character_,
  download_url_chr = NA_character_,
  inc_file_main_chr = NA_character_,
  inc_fls_to_rename_ls = list(),
  new_nms_for_inc_fls_ls = list(),
  filter_cdn_1L_chr = NA_character_,
  local_to_url_vec_chr = NA_character_,
  slice_indcs_int = NA_integer_,
  urls_vec_chr = NA_character_
)

\S4method{renew}{ready4use_imports}(
  x,
  local_file_src_chr = NA_character_,
  path_to_make_script_chr = NA_character_,
  download_url_chr = NA_character_,
  inc_file_main_chr = NA_character_,
  inc_fls_to_rename_ls = list(),
  new_nms_for_inc_fls_ls = list(),
  filter_cdn_1L_chr = NA_character_,
  local_to_url_vec_chr = NA_character_,
  slice_indcs_int = NA_integer_,
  urls_vec_chr = NA_character_
)

\S4method{renew}{Ready4useDyad}(
  x,
  arrange_by_1L_chr = c("category", "name", "both"),
  categories_chr = character(0),
  drop_chr = character(0),
  dictionary_r3 = ready4use_dictionary(),
  dummys_ls = NULL,
  factors_chr = character(0),
  fn = NULL,
  fn_args_ls = NULL,
  names_chr = character(0),
  new_val_xx = NULL,
  remove_old_lbls_1L_lgl = T,
  tfmn_1L_chr = "capitalise",
  type_1L_chr = c("label", "base", "case", "drop", "dummys", "join", "keep", "levels",
    "mutate", "rbind", "unlabel"),
  uid_var_nm_1L_chr = character(0),
  var_ctg_chr = "Uncategorised",
  what_1L_chr = c("all", "dataset", "dictionary"),
  ...
)
}
\arguments{
\item{x}{An object of class Ready4useDyad}

\item{var_nm_chr}{Variable name (a character vector), Default: 'NA'}

\item{var_ctg_chr}{Variable category (a character vector), Default: 'Uncategorised'}

\item{var_desc_chr}{Variable description (a character vector), Default: 'NA'}

\item{var_type_chr}{Variable type (a character vector), Default: 'NA'}

\item{filter_cdn_1L_chr}{Filter condition (a character vector of length one), Default: 'NA'}

\item{new_cases_r3}{New cases (a ready4 submodule), Default: NULL}

\item{new_ready4_dict_r3}{New ready4 dictionary (a ready4 submodule), Default: deprecated()}

\item{slice_indcs_int}{Slice indices (an integer vector), Default: NA}

\item{local_file_src_chr}{Local file source (a character vector), Default: 'NA'}

\item{path_to_make_script_chr}{Path to make script (a character vector), Default: 'NA'}

\item{download_url_chr}{Download url (a character vector), Default: 'NA'}

\item{inc_file_main_chr}{Include file main (a character vector), Default: 'NA'}

\item{inc_fls_to_rename_ls}{Include files to rename (a list), Default: list()}

\item{new_nms_for_inc_fls_ls}{New names for include files (a list), Default: list()}

\item{local_to_url_vec_chr}{Local to url vector (a character vector), Default: 'NA'}

\item{urls_vec_chr}{Urls vector (a character vector), Default: 'NA'}

\item{arrange_by_1L_chr}{Arrange by (a character vector of length one), Default: c("category", "name", "both")}

\item{categories_chr}{Categories (a character vector), Default: character(0)}

\item{drop_chr}{Drop (a character vector), Default: character(0)}

\item{dictionary_r3}{Dictionary (a ready4 submodule), Default: ready4use_dictionary()}

\item{dummys_ls}{Dummys (a list), Default: NULL}

\item{factors_chr}{Factors (a character vector), Default: character(0)}

\item{fn}{Function (a function), Default: NULL}

\item{fn_args_ls}{Function arguments (a list), Default: NULL}

\item{names_chr}{Names (a character vector), Default: character(0)}

\item{new_val_xx}{New value (an output object of multiple potential types), Default: NULL}

\item{remove_old_lbls_1L_lgl}{Remove old labels (a logical vector of length one), Default: T}

\item{tfmn_1L_chr}{Transformation (a character vector of length one), Default: 'capitalise'}

\item{type_1L_chr}{Type (a character vector of length one), Default: c("label", "base", "case", "drop", "dummys", "join", "keep",
"levels", "mutate", "rbind", "unlabel")}

\item{uid_var_nm_1L_chr}{Unique identifier variable name (a character vector of length one), Default: character(0)}

\item{what_1L_chr}{What (a character vector of length one), Default: c("all", "dataset", "dictionary")}

\item{...}{Additional arguments}
}
\value{
x (An object)

x (An object)

x (An object of class Ready4useDyad)
}
\description{
renew.ready4use_dictionary() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 s3 class defining a data dictionary tibble. The function is called for its side effects and does not return a value.

renew.ready4use_imports() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.

renew method applied to Ready4useDyad
}
