% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_generics.R
\name{update_src_loc_to_url}
\alias{update_src_loc_to_url}
\title{Update source local to url}
\usage{
update_src_loc_to_url(x, local_to_url_vec, urls_vec, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{local_to_url_vec}{PARAM_DESCRIPTION}

\item{urls_vec}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\description{
update_src_loc_to_url() is an Update Source Local to Url generic that updates data from a local file reference to a URL
}
\keyword{internal}
