% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_dv_file_fl}
\alias{write_dv_file_fl}
\title{Write dataverse file file}
\usage{
write_dv_file_fl(
  database_ui_chr,
  filename_chr,
  repo_file_format,
  dataverse_chr = Sys.getenv("DATAVERSE_SERVER"),
  save_type_chr = "original",
  destination_path_chr
)
}
\arguments{
\item{database_ui_chr}{Database ui (a character vector of length 1)}

\item{filename_chr}{Filename (a character vector of length 1)}

\item{repo_file_format}{PARAM_DESCRIPTION}

\item{dataverse_chr}{Dataverse (a character vector of length 1), Default: Sys.getenv("DATAVERSE_SERVER")}

\item{save_type_chr}{Save type (a character vector of length 1), Default: 'original'}

\item{destination_path_chr}{Destination path (a character vector of length 1)}
}
\description{
write_dv_file_fl() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write dv a file file.NA
}
\keyword{internal}
